The BetaML.GMM Module

BetaML.GMM — Module

GMM module

Generative (Gaussian) Mixed Model learners (supervised/unsupervised)

Provides clustering and regressors using (Generative) Gaussiam Mixture Model (probabilistic).

Collaborative filtering / missing values imputation / reccomendation systems based on GMM is available in the Imputation module.

The module provides the following models. Use ?[model] to access their documentation:

GMMClusterer: soft-clustering using GMM

GMMRegressor1: regressor using GMM as back-end (first algorithm)

GMMRegressor1: regressor using GMM as back-end (second algorithm)

All the algorithms works with arbitrary mixture distribution, altought only {Spherical|Diagonal|Full} Gaussian mixtures has been implemented. User defined mixtures can be used defining a struct as subtype of AbstractMixture and implementing for that mixture the following functions:

init_mixtures!(mixtures, X; minimum_variance, minimum_covariance, initialisation_strategy)

lpdf(m,x,mask) (for the e-step)

update_parameters!(mixtures, X, pₙₖ; minimum_variance, minimum_covariance) (the m-step)

npar(mixtures::Array{T,1}) (for the BIC/AIC computation)

All the GMM-based algorithms works only with numerical data, but accepts also Missing one.

Module Index

BetaML.GMM.GMMClusterer

BetaML.GMM.GMMHyperParametersSet

BetaML.GMM.GMMRegressor1

BetaML.GMM.GMMRegressor2

BetaML.GMM.GaussianMixtureClusterer

BetaML.GMM.GaussianMixtureRegressor

BetaML.GMM.MultitargetGaussianMixtureRegressor

BetaML.GMM.init_mixtures!

BetaML.GMM.lpdf

BetaML.GMM.lpdf

BetaML.GMM.lpdf

Detailed API

BetaML.GMM.GMMClusterer — Type

mutable struct GMMClusterer <: BetaMLUnsupervisedModel

Assign class probabilities to records (i.e. soft clustering) assuming a probabilistic generative model of observed data using mixtures.

For the parameters see ?GMMHyperParametersSet and ?BetaMLDefaultOptionsSet.

Notes:

Data must be numerical

Mixtures can be user defined: see the ?GMM module documentation for a discussion on provided vs custom mixtures.

Online fitting (re-fitting with new data) is supported by setting the old learned mixtrures as the starting values

The model is fitted using an Expectation-Minimisation (EM) algorithm that supports Missing data and is implemented in the log-domain for better numerical accuracy with many dimensions

Example:

julia> using BetaML

julia> X = [1.1 10.1; 0.9 9.8; 10.0 1.1; 12.1 0.8; 0.8 9.8];

julia> mod = GMMClusterer(n_classes=2)
GMMClusterer - A Generative Mixture Model (unfitted)

julia> prob_belong_classes = fit!(mod,X)
Iter. 1:        Var. of the post  2.15612140465882        Log-likelihood -29.06452054772657
5×2 Matrix{Float64}:
 1.0  0.0
 1.0  0.0
 0.0  1.0
 0.0  1.0
 1.0  0.0

julia> new_probs = fit!(mod,[11 0.9])
Iter. 1:        Var. of the post  1.0     Log-likelihood -1.3312256125240092
1×2 Matrix{Float64}:
 0.0  1.0

julia> info(mod)
Dict{String, Any} with 6 entries:
  "xndims"         => 2
  "error"          => [1.0, 0.0, 0.0]
  "AIC"            => 15.7843
  "fitted_records" => 6
  "lL"             => 1.10786
  "BIC"            => -2.21571

julia> parameters(mod)
BetaML.GMM.GMMClusterLearnableParameters (a BetaMLLearnableParametersSet struct)
- mixtures: DiagonalGaussian{Float64}[DiagonalGaussian{Float64}([0.9333333333333332, 9.9], [0.05, 0.05]), DiagonalGaussian{Float64}([11.05, 0.9500000000000001], [0.05, 0.05])]
- initial_probmixtures: [0.0, 1.0]

BetaML.GMM.GMMHyperParametersSet — Type

mutable struct GMMHyperParametersSet <: BetaMLHyperParametersSet

Hyperparameters for GMM clusters and other GMM-based algorithms

Parameters:

n_classes

: Number of mixtures (latent classes) to consider [def: 3]

initial_probmixtures

: Initial probabilities of the categorical distribution (n_classes x 1) [default: []]

mixtures

: An array (of length n_classes) of the mixtures to employ (see the [?GMM](@ref GMM) module). Each mixture object can be provided with or without its parameters (e.g. mean and variance for the gaussian ones). Fully qualified mixtures are useful only if theinitialisationstrategyparameter is set to "gived" This parameter can also be given symply in term of a _type. In this case it is automatically extended to a vector of n_classesmixtures of the specified type. Note that mixing of different mixture types is not currently supported and that currently implemented mixtures areSphericalGaussian,DiagonalGaussianandFullGaussian. [def:DiagonalGaussian`]

tol

: Tolerance to stop the algorithm [default: 10^(-6)]

minimum_variance

: Minimum variance for the mixtures [default: 0.05]

minimum_covariance

: Minimum covariance for the mixtures with full covariance matrix [default: 0]. This should be set different than minimum_variance.

initialisation_strategy

: The computation method of the vector of the initial mixtures. One of the following: - "grid": using a grid approach - "given": using the mixture provided in the fully qualified mixtures parameter - "kmeans": use first kmeans (itself initialised with a "grid" strategy) to set the initial mixture centers [default] Note that currently "random" and "shuffle" initialisations are not supported in gmm-based algorithms.

maximum_iterations

: Maximum number of iterations [def: 5000]

tunemethod

: The method - and its parameters - to employ for hyperparameters autotuning. See `SuccessiveHalvingSearch for the default method (suitable for the GMM-based regressors) To implement automatic hyperparameter tuning during the (first) fit! call simply set autotune=true and eventually change the default tunemethod options (including the parameter ranges, the resources to employ and the loss function to adopt).

BetaML.GMM.GMMRegressor1 — Type

mutable struct GMMRegressor1 <: BetaMLUnsupervisedModel

A multi-dimensional, missing data friendly non-linear regressor based on Generative (Gaussian) Mixture Model (strategy "1").

The training data is used to fit a probabilistic model with latent mixtures (Gaussian distributions with different covariances are already implemented) and then predictions of new data is obtained by fitting the new data to the mixtures.

For hyperparameters see GMMHyperParametersSet and BetaMLDefaultOptionsSet.

This strategy (GMMRegressor1) works by fitting the EM algorithm on the feature matrix X. Once the data has been probabilistically assigned to the various classes, a mean value of fitting values Y is computed for each cluster (using the probabilities as weigths). At predict time, the new data is first fitted to the learned mixtures using the e-step part of the EM algorithm to obtain the probabilistic assignment of each record to the various mixtures. Then these probabilities are multiplied to the mixture averages for the Y dimensions learned at training time to obtain the predicted value(s) for each record.

Notes:

Predicted values are always a matrix, even when a single variable is predicted (use dropdims(ŷ,dims=2) to get a single vector).

Example:

julia> using BetaML

julia> X = [1.1 10.1; 0.9 9.8; 10.0 1.1; 12.1 0.8; 0.8 9.8];

julia> Y = X[:,1] .* 2 - X[:,2]
5-element Vector{Float64}:
 -7.8999999999999995
 -8.0
 18.9
 23.4
 -8.200000000000001

julia> mod = GMMRegressor1(n_classes=2)
GMMRegressor1 - A regressor based on Generative Mixture Model (unfitted)

julia> ŷ = fit!(mod,X,Y)
Iter. 1:        Var. of the post  2.15612140465882        Log-likelihood -29.06452054772657
5×1 Matrix{Float64}:
 -8.033333333333333
 -8.033333333333333
 21.15
 21.15
 -8.033333333333333

julia> new_probs = predict(mod,[11 0.9])
1×1 Matrix{Float64}:
 21.15

julia> info(mod)
Dict{String, Any} with 6 entries:
  "xndims"         => 2
  "error"          => [2.15612, 0.118848, 4.19495e-7, 0.0, 0.0]
  "AIC"            => 32.7605
  "fitted_records" => 5
  "lL"             => -7.38023
  "BIC"            => 29.2454

BetaML.GMM.GMMRegressor2 — Type

mutable struct GMMRegressor2 <: BetaMLUnsupervisedModel

A multi-dimensional, missing data friendly non-linear regressor based on Generative (Gaussian) Mixture Model.

The training data is used to fit a probabilistic model with latent mixtures (Gaussian distributions with different covariances are already implemented) and then predictions of new data is obtained by fitting the new data to the mixtures.

For hyperparameters see GMMHyperParametersSet and BetaMLDefaultOptionsSet.

Thsi strategy (GMMRegressor2) works by training the EM algorithm on a combined (hcat) matrix of X and Y. At predict time, the new data is first fitted to the learned mixtures using the e-step part of the EM algorithm (and using missing values for the dimensions belonging to Y) to obtain the probabilistic assignment of each record to the various mixtures. Then these probabilities are multiplied to the mixture averages for the Y dimensions to obtain the predicted value(s) for each record.

Example:

julia> using BetaML

julia> X = [1.1 10.1; 0.9 9.8; 10.0 1.1; 12.1 0.8; 0.8 9.8];

julia> Y = X[:,1] .* 2 - X[:,2]
5-element Vector{Float64}:
 -7.8999999999999995
 -8.0
 18.9
 23.4
 -8.200000000000001

julia> mod = GMMRegressor2(n_classes=2)
GMMRegressor2 - A regressor based on Generative Mixture Model (unfitted)

julia> ŷ = fit!(mod,X,Y)
Iter. 1:        Var. of the post  2.2191120060614065      Log-likelihood -47.70971887023561
5×1 Matrix{Float64}:
 -8.033333333333333
 -8.033333333333333
 21.15
 21.15
 -8.033333333333333

julia> new_probs = predict(mod,[11 0.9])
1×1 Matrix{Float64}:
 21.15

julia> info(mod)
Dict{String, Any} with 6 entries:
  "xndims"         => 3
  "error"          => [2.21911, 0.0260833, 3.19141e-39, 0.0]
  "AIC"            => 60.0684
  "fitted_records" => 5
  "lL"             => -17.0342
  "BIC"            => 54.9911

julia> parameters(mod)
BetaML.GMM.GMMClusterLearnableParameters (a BetaMLLearnableParametersSet struct)
- mixtures: DiagonalGaussian{Float64}[DiagonalGaussian{Float64}([0.9333333333333332, 9.9, -8.033333333333333], [1.1024999999999996, 0.05, 5.0625]), DiagonalGaussian{Float64}([11.05, 0.9500000000000001, 21.15], [1.1024999999999996, 0.05, 5.0625])]
- initial_probmixtures: [0.6, 0.4]

BetaML.GMM.GaussianMixtureClusterer — Type

mutable struct GaussianMixtureClusterer <: MLJModelInterface.Unsupervised

A Expectation-Maximisation clustering algorithm with customisable mixtures, from the Beta Machine Learning Toolkit (BetaML).

Hyperparameters:

n_classes::Int64

: Number of mixtures (latent classes) to consider [def: 3]

initial_probmixtures::AbstractVector{Float64}

: Initial probabilities of the categorical distribution (n_classes x 1) [default: []]

mixtures::Union{Type, Vector{var"#s792"} where var"#s792"<:AbstractMixture}

: An array (of length n_classes) of the mixtures to employ (see the [?GMM](@ref GMM) module). Each mixture object can be provided with or without its parameters (e.g. mean and variance for the gaussian ones). Fully qualified mixtures are useful only if theinitialisationstrategyparameter is set to "gived" This parameter can also be given symply in term of a _type. In this case it is automatically extended to a vector of n_classesmixtures of the specified type. Note that mixing of different mixture types is not currently supported. [def:[DiagonalGaussian() for i in 1:n_classes]`]

tol::Float64

: Tolerance to stop the algorithm [default: 10^(-6)]

minimum_variance::Float64

: Minimum variance for the mixtures [default: 0.05]

minimum_covariance::Float64

: Minimum covariance for the mixtures with full covariance matrix [default: 0]. This should be set different than minimum_variance (see notes).

initialisation_strategy::String

: The computation method of the vector of the initial mixtures. One of the following: - "grid": using a grid approach - "given": using the mixture provided in the fully qualified mixtures parameter - "kmeans": use first kmeans (itself initialised with a "grid" strategy) to set the initial mixture centers [default] Note that currently "random" and "shuffle" initialisations are not supported in gmm-based algorithms.

maximum_iterations::Int64

: Maximum number of iterations [def: typemax(Int64), i.e. ∞]

rng::Random.AbstractRNG

: Random Number Generator [deafult: Random.GLOBAL_RNG]

Example:

julia> using MLJ

julia> modelType                   = @load GaussianMixtureClusterer pkg = "BetaML"
[ Info: For silent loading, specify `verbosity=0`. 
import BetaML ✔
BetaML.GMM.GaussianMixtureClusterer

julia> model                       = modelType()
GaussianMixtureClusterer(
  n_classes = 3, 
  initial_probmixtures = Float64[], 
  mixtures = BetaML.GMM.DiagonalGaussian{Float64}[BetaML.GMM.DiagonalGaussian{Float64}(nothing, nothing), BetaML.GMM.DiagonalGaussian{Float64}(nothing, nothing), BetaML.GMM.DiagonalGaussian{Float64}(nothing, nothing)], 
  tol = 1.0e-6, 
  minimum_variance = 0.05, 
  minimum_covariance = 0.0, 
  initialisation_strategy = "kmeans", 
  maximum_iterations = 9223372036854775807, 
  rng = Random._GLOBAL_RNG())

julia> X, y                        = @load_iris;

julia> (fitResults, cache, report) = MLJ.fit(model, 0, X);

julia> est_classes                 = predict(model, fitResults, X)
150-element CategoricalDistributions.UnivariateFiniteVector{Multiclass{3}, Int64, UInt32, Float64}:
 UnivariateFinite{Multiclass{3}}(1=>1.0, 2=>4.17e-15, 3=>2.1900000000000003e-31)
 UnivariateFinite{Multiclass{3}}(1=>1.0, 2=>1.25e-13, 3=>5.87e-31)
 UnivariateFinite{Multiclass{3}}(1=>1.0, 2=>4.5e-15, 3=>1.55e-32)
 ⋮
 UnivariateFinite{Multiclass{3}}(1=>5.39e-25, 2=>0.0167, 3=>0.983)
 UnivariateFinite{Multiclass{3}}(1=>7.5e-29, 2=>0.000106, 3=>1.0)
 UnivariateFinite{Multiclass{3}}(1=>1.6e-20, 2=>0.594, 3=>0.406)

julia> 

BetaML.GMM.GaussianMixtureRegressor — Type

mutable struct GaussianMixtureRegressor <: MLJModelInterface.Deterministic

A non-linear regressor derived from fitting the data on a probabilistic model (Gaussian Mixture Model). Relatively fast.

This is the single-target version of the model. If you want to predict several labels (y) at once, use the MLJ model MultitargetGaussianMixtureRegressor.

Hyperparameters:

n_classes::Int64

: Number of mixtures (latent classes) to consider [def: 3]

initial_probmixtures::Vector{Float64}

: Initial probabilities of the categorical distribution (n_classes x 1) [default: []]

mixtures::Union{Type, Vector{var"#s792"} where var"#s792"<:AbstractMixture}

: An array (of length n_classes) of the mixtures to employ (see the [?GMM](@ref GMM) module). Each mixture object can be provided with or without its parameters (e.g. mean and variance for the gaussian ones). Fully qualified mixtures are useful only if theinitialisationstrategyparameter is set to "gived" This parameter can also be given symply in term of a _type. In this case it is automatically extended to a vector of n_classesmixtures of the specified type. Note that mixing of different mixture types is not currently supported. [def:[DiagonalGaussian() for i in 1:n_classes]`]

tol::Float64

: Tolerance to stop the algorithm [default: 10^(-6)]

minimum_variance::Float64

: Minimum variance for the mixtures [default: 0.05]

minimum_covariance::Float64

: Minimum covariance for the mixtures with full covariance matrix [default: 0]. This should be set different than minimum_variance (see notes).

initialisation_strategy::String

: The computation method of the vector of the initial mixtures. One of the following: - "grid": using a grid approach - "given": using the mixture provided in the fully qualified mixtures parameter - "kmeans": use first kmeans (itself initialised with a "grid" strategy) to set the initial mixture centers [default] Note that currently "random" and "shuffle" initialisations are not supported in gmm-based algorithms.

maximum_iterations::Int64

: Maximum number of iterations [def: typemax(Int64), i.e. ∞]

rng::Random.AbstractRNG

: Random Number Generator [deafult: Random.GLOBAL_RNG]

Example:

julia> using MLJ

julia> modelType                   = @load GaussianMixtureRegressor pkg = "BetaML"
[ Info: For silent loading, specify `verbosity=0`. 
import BetaML ✔
BetaML.GMM.GaussianMixtureRegressor

julia> model                       = modelType()
GaussianMixtureRegressor(
n_classes = 3, 
initial_probmixtures = Float64[], 
mixtures = BetaML.GMM.DiagonalGaussian{Float64}[BetaML.GMM.DiagonalGaussian{Float64}(nothing, nothing), BetaML.GMM.DiagonalGaussian{Float64}(nothing, nothing), BetaML.GMM.DiagonalGaussian{Float64}(nothing, nothing)], 
tol = 1.0e-6, 
minimum_variance = 0.05, 
minimum_covariance = 0.0, 
initialisation_strategy = "kmeans", 
maximum_iterations = 9223372036854775807, 
rng = Random._GLOBAL_RNG())

julia> X, y                        = @load_boston;

julia> (fitResults, cache, report) = MLJ.fit(model, 0, X, y);

julia> y_est                       = predict(model, fitResults, X)
506-element Vector{Float64}:
24.703442835305577
24.70344283512716
24.70344283528249
⋮
17.172486989759676
17.172486989759676
17.172486989759644

BetaML.GMM.MultitargetGaussianMixtureRegressor — Type

mutable struct MultitargetGaussianMixtureRegressor <: MLJModelInterface.Deterministic

A non-linear regressor derived from fitting the data on a probabilistic model (Gaussian Mixture Model). Relatively fast.

This is the multi-target version of the model. If you want to predict a single label (y), use the MLJ model GaussianMixtureRegressor.

Hyperparameters:

n_classes::Int64

: Number of mixtures (latent classes) to consider [def: 3]

initial_probmixtures::Vector{Float64}

: Initial probabilities of the categorical distribution (n_classes x 1) [default: []]

mixtures::Union{Type, Vector{var"#s792"} where var"#s792"<:AbstractMixture}

: An array (of length n_classes) of the mixtures to employ (see the [?GMM](@ref GMM) module). Each mixture object can be provided with or without its parameters (e.g. mean and variance for the gaussian ones). Fully qualified mixtures are useful only if theinitialisationstrategyparameter is set to "gived" This parameter can also be given symply in term of a _type. In this case it is automatically extended to a vector of n_classesmixtures of the specified type. Note that mixing of different mixture types is not currently supported. [def:[DiagonalGaussian() for i in 1:n_classes]`]

tol::Float64

: Tolerance to stop the algorithm [default: 10^(-6)]

minimum_variance::Float64

: Minimum variance for the mixtures [default: 0.05]

minimum_covariance::Float64

: Minimum covariance for the mixtures with full covariance matrix [default: 0]. This should be set different than minimum_variance (see notes).

initialisation_strategy::String

: The computation method of the vector of the initial mixtures. One of the following: - "grid": using a grid approach - "given": using the mixture provided in the fully qualified mixtures parameter - "kmeans": use first kmeans (itself initialised with a "grid" strategy) to set the initial mixture centers [default] Note that currently "random" and "shuffle" initialisations are not supported in gmm-based algorithms.

maximum_iterations::Int64

: Maximum number of iterations [def: typemax(Int64), i.e. ∞]

rng::Random.AbstractRNG

: Random Number Generator [deafult: Random.GLOBAL_RNG]

Example:

julia> using MLJ

julia> modelType                   = @load MultitargetGaussianMixtureRegressor pkg = "BetaML"
[ Info: For silent loading, specify `verbosity=0`. 
import BetaML ✔
BetaML.GMM.MultitargetGaussianMixtureRegressor
julia> model                       = modelType()
MultitargetGaussianMixtureRegressor(
  n_classes = 3, 
  initial_probmixtures = Float64[], 
  mixtures = BetaML.GMM.DiagonalGaussian{Float64}[BetaML.GMM.DiagonalGaussian{Float64}(nothing, nothing), BetaML.GMM.DiagonalGaussian{Float64}(nothing, nothing), BetaML.GMM.DiagonalGaussian{Float64}(nothing, nothing)], 
  tol = 1.0e-6, 
  minimum_variance = 0.05, 
  minimum_covariance = 0.0, 
  initialisation_strategy = "kmeans", 
  maximum_iterations = 9223372036854775807, 
  rng = Random._GLOBAL_RNG())

julia> X, y                        = @load_boston;

julia> ydouble = hcat(y,y);

julia> (fitResults, cache, report) = MLJ.fit(model, 0, X, ydouble);

julia> y_est                       = predict(model, fitResults, X)
506×2 Matrix{Float64}:
 24.5785  24.5785
 24.5785  24.5785
 24.5785  24.5785
  ⋮       
 17.0039  17.0039
 17.0039  17.0039
 17.0039  17.0039

BetaML.GMM.init_mixtures! — Method

init_mixtures!(mixtures::Array{T,1}, X; minimum_variance=0.25, minimum_covariance=0.0, initialisation_strategy="grid",rng=Random.GLOBAL_RNG)

The parameter initialisation_strategy can be grid, kmeans or given:

grid: Uniformly cover the space observed by the data

kmeans: Use the kmeans algorithm. If the data contains missing values, a first run of predictMissing is done under init=grid to impute the missing values just to allow the kmeans algorithm. Then the em algorithm is used with the output of kmean as init values.

given: Leave the provided set of initial mixtures

BetaML.GMM.lpdf — Method

lpdf(m::DiagonalGaussian,x,mask) - Log PDF of the mixture given the observation x

BetaML.GMM.lpdf — Method

lpdf(m::FullGaussian,x,mask) - Log PDF of the mixture given the observation x

BetaML.GMM.lpdf — Method

lpdf(m::SphericalGaussian,x,mask) - Log PDF of the mixture given the observation x