The BetaML.Perceptron Module

BetaML.Perceptron — Module

Perceptron module

Provide linear and kernel classifiers.

Provide the following supervised models:

PerceptronClassifier: Train data using the classical perceptron

KernelPerceptronClassifier: Train data using the kernel perceptron

PegasosClassifier: Train data using the pegasos algorithm

All algorithms are multiclass, with PerceptronClassifier and PegasosClassifier employing a one-vs-all strategy, while KernelPerceptronClassifier employs a one-vs-one approach, and return a "probability" for each class in term of a dictionary for each record. Use mode(ŷ) to return a single class prediction per record.

These models are available in the MLJ framework as LinearPerceptron,KernelPerceptron and Pegasos respectivly.

Module Index

BetaML.Perceptron.KernelPerceptron

BetaML.Perceptron.KernelPerceptronClassifier

BetaML.Perceptron.KernelPerceptronClassifierHyperParametersSet

BetaML.Perceptron.LinearPerceptron

BetaML.Perceptron.Pegasos

BetaML.Perceptron.PegasosClassifier

BetaML.Perceptron.PegasosClassifierHyperParametersSet

BetaML.Perceptron.PerceptronClassifier

BetaML.Perceptron.PerceptronClassifierHyperParametersSet

Detailed API

BetaML.Perceptron.KernelPerceptron — Type

mutable struct KernelPerceptron <: MLJModelInterface.Probabilistic

The kernel perceptron algorithm using one-vs-one for multiclass, from the Beta Machine Learning Toolkit (BetaML).

Hyperparameters:

kernel::Function

: Kernel function to employ. See ?radial_kernel or ?polynomial_kernel (once loaded the BetaML package) for details or check ?BetaML.Utils to verify if other kernels are defined (you can alsways define your own kernel) [def: radial_kernel]

epochs::Int64

: Maximum number of epochs, i.e. passages trough the whole training sample [def: 100]

initial_errors::Union{Nothing, Vector{Vector{Int64}}}

: Initial distribution of the number of errors errors [def: nothing, i.e. zeros]. If provided, this should be a nModels-lenght vector of nRecords integer values vectors , where nModels is computed as (n_classes * (n_classes - 1)) / 2

shuffle::Bool

: Whether to randomly shuffle the data at each iteration (epoch) [def: true]

rng::Random.AbstractRNG

: A Random Number Generator to be used in stochastic parts of the code [deafult: Random.GLOBAL_RNG]

BetaML.Perceptron.KernelPerceptronClassifier — Type

mutable struct KernelPerceptronClassifier <: BetaMLSupervisedModel

A "kernel" version of the Perceptron model (supervised) with user configurable kernel function.

For the parameters see ?KernelPerceptronClassifierHyperParametersSet and ?BetaMLDefaultOptionsSet

Limitations:

data must be numerical

online training (retraining) is not supported

BetaML.Perceptron.KernelPerceptronClassifierHyperParametersSet — Type

mutable struct KernelPerceptronClassifierHyperParametersSet <: BetaMLHyperParametersSet

Hyperparameters for the KernelPerceptronClassifier model

Parameters:

kernel

: Kernel function to employ. See ?radial_kernel or ?polynomial_kernel for details or check ?BetaML.Utils to verify if other kernels are defined (you can alsways define your own kernel) [def: radial_kernel]

initial_errors

: Initial distribution of the number of errors errors [def: nothing, i.e. zeros]. If provided, this should be a nModels-lenght vector of nRecords integer values vectors , where nModels is computed as (n_classes * (n_classes - 1)) / 2

epochs

: Maximum number of epochs, i.e. passages trough the whole training sample [def: 100]

shuffle

: Whether to randomly shuffle the data at each iteration (epoch) [def: true]

tunemethod

: The method - and its parameters - to employ for hyperparameters autotuning. See SuccessiveHalvingSearch for the default method. To implement automatic hyperparameter tuning during the (first) fit! call simply set autotune=true and eventually change the default tunemethod options (including the parameter ranges, the resources to employ and the loss function to adopt).

BetaML.Perceptron.LinearPerceptron — Type

mutable struct LinearPerceptron <: MLJModelInterface.Probabilistic

The classical perceptron algorithm using one-vs-all for multiclass, from the Beta Machine Learning Toolkit (BetaML).

Hyperparameters:

initial_coefficients::Union{Nothing, Matrix{Float64}}

: N-classes by D-dimensions matrix of initial linear coefficients [def: nothing, i.e. zeros]

initial_constant::Union{Nothing, Vector{Float64}}

: N-classes vector of initial contant terms [def: nothing, i.e. zeros]

epochs::Int64

: Maximum number of epochs, i.e. passages trough the whole training sample [def: 1000]

shuffle::Bool

: Whether to randomly shuffle the data at each iteration (epoch) [def: true]

force_origin::Bool

: Whether to force the parameter associated with the constant term to remain zero [def: false]

return_mean_hyperplane::Bool

: Whether to return the average hyperplane coefficients instead of the final ones [def: false]

rng::Random.AbstractRNG

: A Random Number Generator to be used in stochastic parts of the code [deafult: Random.GLOBAL_RNG]

BetaML.Perceptron.Pegasos — Type

mutable struct Pegasos <: MLJModelInterface.Probabilistic

The gradient-based linear "pegasos" classifier using one-vs-all for multiclass, from the Beta Machine Learning Toolkit (BetaML).

Hyperparameters:

initial_coefficients::Union{Nothing, Matrix{Float64}}

: N-classes by D-dimensions matrix of initial linear coefficients [def: nothing, i.e. zeros]

initial_constant::Union{Nothing, Vector{Float64}}

: N-classes vector of initial contant terms [def: nothing, i.e. zeros]

learning_rate::Function

: Learning rate [def: (epoch -> 1/sqrt(epoch))]

learning_rate_multiplicative::Float64

: Multiplicative term of the learning rate [def: 0.5]

epochs::Int64

: Maximum number of epochs, i.e. passages trough the whole training sample [def: 1000]

shuffle::Bool

: Whether to randomly shuffle the data at each iteration (epoch) [def: true]

force_origin::Bool

: Whether to force the parameter associated with the constant term to remain zero [def: false]

return_mean_hyperplane::Bool

: Whether to return the average hyperplane coefficients instead of the final ones [def: false]

rng::Random.AbstractRNG

: A Random Number Generator to be used in stochastic parts of the code [deafult: Random.GLOBAL_RNG]

BetaML.Perceptron.PegasosClassifier — Type

mutable struct PegasosClassifier <: BetaMLSupervisedModel

The PegasosClassifier model, a linear, gradient-based classifier. Multiclass is supported using a one-vs-all approach.

See ?PegasosClassifierHyperParametersSet and ?BetaMLDefaultOptionsSet for applicable hyperparameters and options.

BetaML.Perceptron.PegasosClassifierHyperParametersSet — Type

mutable struct PegasosClassifierHyperParametersSet <: BetaMLHyperParametersSet

Hyperparameters for the PegasosClassifier model.

Parameters:

learning_rate::Function

: Learning rate [def: (epoch -> 1/sqrt(epoch))]

learning_rate_multiplicative::Float64

: Multiplicative term of the learning rate [def: 0.5]

initial_parameters::Union{Nothing, Matrix{Float64}}

: Initial parameters. If given, should be a matrix of n-classes by feature dimension + 1 (to include the constant term as the first element) [def: nothing, i.e. zeros]

epochs::Int64

: Maximum number of epochs, i.e. passages trough the whole training sample [def: 1000]

shuffle::Bool

: Whether to randomly shuffle the data at each iteration (epoch) [def: true]

force_origin::Bool

: Whether to force the parameter associated with the constant term to remain zero [def: false]

return_mean_hyperplane::Bool

: Whether to return the average hyperplane coefficients instead of the final ones [def: false]

tunemethod::AutoTuneMethod

: The method - and its parameters - to employ for hyperparameters autotuning. See SuccessiveHalvingSearch for the default method. To implement automatic hyperparameter tuning during the (first) fit! call simply set autotune=true and eventually change the default tunemethod options (including the parameter ranges, the resources to employ and the loss function to adopt).

BetaML.Perceptron.PerceptronClassifier — Type

mutable struct PerceptronClassifier <: BetaMLSupervisedModel

The classical "perceptron" linear classifier (supervised).

For the parameters see ?PerceptronClassifierHyperParametersSet and ?BetaMLDefaultOptionsSet.

Notes:

data must be numerical

online fitting (re-fitting with new data) is not supported

BetaML.Perceptron.PerceptronClassifierHyperParametersSet — Type

mutable struct PerceptronClassifierHyperParametersSet <: BetaMLHyperParametersSet

Hyperparameters for the PerceptronClassifier model

Parameters:

initial_parameters::Union{Nothing, Matrix{Float64}}

: Initial parameters. If given, should be a matrix of n-classes by feature dimension + 1 (to include the constant term as the first element) [def: nothing, i.e. zeros]

epochs::Int64

: Maximum number of epochs, i.e. passages trough the whole training sample [def: 1000]

shuffle::Bool

: Whether to randomly shuffle the data at each iteration (epoch) [def: true]

force_origin::Bool

: Whether to force the parameter associated with the constant term to remain zero [def: false]

return_mean_hyperplane::Bool

: Whether to return the average hyperplane coefficients instead of the final ones [def: false]

tunemethod::AutoTuneMethod

: The method - and its parameters - to employ for hyperparameters autotuning. See SuccessiveHalvingSearch for the default method. To implement automatic hyperparameter tuning during the (first) fit! call simply set autotune=true and eventually change the default tunemethod options (including the parameter ranges, the resources to employ and the loss function to adopt).