The BetaML.Utils Module

BetaML.Utils — Module

Utils module

Provide shared utility functions for various machine learning algorithms. You don't usually need to import from this module, as each other module (Nn, Perceptron, Clusters,...) reexport it.

Module Index

BetaML.Utils.FIXEDSEED

BetaML.Utils.accuracy

BetaML.Utils.accuracy

BetaML.Utils.accuracy

BetaML.Utils.accuracy

BetaML.Utils.accuracy

BetaML.Utils.aic

BetaML.Utils.autoJacobian

BetaML.Utils.batch

BetaML.Utils.bic

BetaML.Utils.celu

BetaML.Utils.classCounts

BetaML.Utils.colsWithMissing

BetaML.Utils.cosine_distance

BetaML.Utils.crossEntropy

BetaML.Utils.dcelu

BetaML.Utils.delu

BetaML.Utils.dmish

BetaML.Utils.dplu

BetaML.Utils.drelu

BetaML.Utils.dsigmoid

BetaML.Utils.dsoftmax

BetaML.Utils.dsoftplus

BetaML.Utils.dtanh

BetaML.Utils.elu

BetaML.Utils.entropy

BetaML.Utils.generateParallelRngs

BetaML.Utils.getPermutations

BetaML.Utils.getScaleFactors

BetaML.Utils.gini

BetaML.Utils.integerDecoder

BetaML.Utils.integerEncoder

BetaML.Utils.issortable

BetaML.Utils.l1_distance

BetaML.Utils.l2_distance

BetaML.Utils.l2²_distance

BetaML.Utils.lse

BetaML.Utils.makeMatrix

BetaML.Utils.meanDicts

BetaML.Utils.meanRelError

BetaML.Utils.mish

BetaML.Utils.mode

BetaML.Utils.oneHotEncoder

BetaML.Utils.pca

BetaML.Utils.plu

BetaML.Utils.polynomialKernel

BetaML.Utils.radialKernel

BetaML.Utils.relu

BetaML.Utils.scale

BetaML.Utils.sigmoid

BetaML.Utils.singleUnique

BetaML.Utils.softmax

BetaML.Utils.softplus

BetaML.Utils.squaredCost

BetaML.Utils.sterling

BetaML.Utils.variance

Detailed API

BetaML.Utils.FIXEDSEED — Constant

FIXEDSEED

Fixed seed to allow reproducible results. This is the seed used to obtain the same results under unit tests.

Use it with:

myAlgorithm(;rng=FIXEDRNG) # always produce the same sequence of results on each run of the script ("pulling" from the same rng object on different calls)

myAlgorithm(;rng=copy(FIXEDRNG) # always produce the same result (new rng object on each call)

Base.error — Method

error(ŷ,y;ignoreLabels=false) - Categorical error (T vs T)

Base.error — Method

error(ŷ,y) - Categorical error with with probabilistic predictions of a dataset given in terms of a dictionary of probabilities (Dict{T,Float64} vs T).

Base.error — Method

error(ŷ,y) - Categorical error with probabilistic predictions of a dataset (PMF vs Int).

Base.error — Method

error(ŷ,y) - Categorical error with probabilistic prediction of a single datapoint (PMF vs Int).

Base.reshape — Method

reshape(myNumber, dims..) - Reshape a number as a n dimensional Array

BetaML.Api.partition — Method

partition(data,parts;shuffle=true)

Partition (by rows) one or more matrices according to the shares in parts.

Parameters

data: A matrix/vector or a vector of matrices/vectors

parts: A vector of the required shares (must sum to 1)

shufle: Whether to randomly shuffle the matrices (preserving the relative order between matrices)

rng: Random Number Generator (see FIXEDSEED) [deafult: Random.GLOBAL_RNG]

Example:

julia julia> x = [1:10 11:20] julia> y = collect(31:40) julia> ((xtrain,xtest),(ytrain,ytest)) = partition([x,y],[0.7,0.3])

BetaML.Utils.accuracy — Method

accuracy(ŷ,y;ignoreLabels=false) - Categorical accuracy between two vectors (T vs T). If

BetaML.Utils.accuracy — Method

accuracy(ŷ,y;tol)

Categorical accuracy with probabilistic predictions of a dataset given in terms of a dictionary of probabilities (Dict{T,Float64} vs T).

Parameters:

ŷ: A narray where each item is the estimated probability mass function in terms of a Dictionary(Item1 => Prob1, Item2 => Prob2, ...)

y: The N array with the correct category for each point $n$.

tol: The tollerance to the prediction, i.e. if considering "correct" only a prediction where the value with highest probability is the true value (tol = 1), or consider instead the set of tol maximum values [def: 1].

BetaML.Utils.accuracy — Method

accuracy(ŷ,y;tol)

Categorical accuracy with probabilistic prediction of a single datapoint given in terms of a dictionary of probabilities (Dict{T,Float64} vs T).

Parameters:

ŷ: The returned probability mass function in terms of a Dictionary(Item1 => Prob1, Item2 => Prob2, ...)

tol: The tollerance to the prediction, i.e. if considering "correct" only a prediction where the value with highest probability is the true value (tol = 1), or consider instead the set of tol maximum values [def: 1].

BetaML.Utils.accuracy — Method

accuracy(ŷ,y;tol,ignoreLabels)

Categorical accuracy with probabilistic predictions of a dataset (PMF vs Int).

Parameters:

ŷ: An (N,K) matrix of probabilities that each $\hat y_n$ record with $n \in 1,....,N$ being of category $k$ with $k \in 1,...,K$.

y: The N array with the correct category for each point $n$.

tol: The tollerance to the prediction, i.e. if considering "correct" only a prediction where the value with highest probability is the true value (tol = 1), or consider instead the set of tol maximum values [def: 1].

ignoreLabels: Whether to ignore the specific label order in y. Useful for unsupervised learning algorithms where the specific label order don't make sense [def: false]

BetaML.Utils.accuracy — Method

accuracy(ŷ,y;tol)

Categorical accuracy with probabilistic prediction of a single datapoint (PMF vs Int).

Use the parameter tol [def: 1] to determine the tollerance of the prediction, i.e. if considering "correct" only a prediction where the value with highest probability is the true value (tol = 1), or consider instead the set of tol maximum values.

BetaML.Utils.aic — Method

aic(lL,k) - Akaike information criterion (lower is better)

BetaML.Utils.autoJacobian — Method

autoJacobian(f,x;nY)

Evaluate the Jacobian using AD in the form of a (nY,nX) madrix of first derivatives

Parameters:

f: The function to compute the Jacobian

x: The input to the function where the jacobian has to be computed

nY: The number of outputs of the function f [def: length(f(x))]

Return values:

An Array{Float64,2} of the locally evaluated Jacobian

Notes:

The nY parameter is optional. If provided it avoids having to compute f(x)

BetaML.Utils.batch — Method

batch(n,bSize;sequential=false,rng)

Return a vector of bSize vectors of indeces from 1 to n. Randomly unless the optional parameter sequential is used.

Example:

julia julia> Utils.batch(6,2,sequential=true) 3-element Array{Array{Int64,1},1}: [1, 2] [3, 4] [5, 6]

BetaML.Utils.bic — Method

bic(lL,k,n) - Bayesian information criterion (lower is better)

BetaML.Utils.celu — Method

celu(x; α=1)

https://arxiv.org/pdf/1704.07483.pdf

BetaML.Utils.classCounts — Method

classCounts(x)

Return a dictionary that counts the number of each unique item (rows) in a dataset.

BetaML.Utils.colsWithMissing — Method

colsWithMissing(x)

Retuyrn an array with the ids of the columns where there is at least a missing value.

BetaML.Utils.cosine_distance — Method

Cosine distance

BetaML.Utils.crossEntropy — Method

crossEntropy(ŷ, y; weight)

Compute the (weighted) cross-entropy between the predicted and the sampled probability distributions.

To be used in classification problems.

BetaML.Utils.dcelu — Method

dcelu(x; α=1)

https://arxiv.org/pdf/1704.07483.pdf

BetaML.Utils.delu — Method

delu(x; α=1) with α > 0

https://arxiv.org/pdf/1511.07289.pdf

BetaML.Utils.dmish — Method

dmish(x)

https://arxiv.org/pdf/1908.08681v1.pdf

BetaML.Utils.dplu — Method

dplu(x;α=0.1,c=1)

Piecewise Linear Unit derivative

https://arxiv.org/pdf/1809.09534.pdf

BetaML.Utils.drelu — Method

drelu(x)

Rectified Linear Unit

https://www.cs.toronto.edu/~hinton/absps/reluICML.pdf

BetaML.Utils.dsigmoid — Method

dsigmoid(x)

BetaML.Utils.dsoftmax — Method

dsoftmax(x; β=1)

Derivative of the softmax function

https://eli.thegreenplace.net/2016/the-softmax-function-and-its-derivative/

BetaML.Utils.dsoftplus — Method

dsoftplus(x)

https://en.wikipedia.org/wiki/Rectifier(neuralnetworks)#Softplus

BetaML.Utils.dtanh — Method

dtanh(x)

BetaML.Utils.elu — Method

elu(x; α=1) with α > 0

https://arxiv.org/pdf/1511.07289.pdf

BetaML.Utils.entropy — Method

entropy(x)

Calculate the entropy for a list of items (or rows).

See: https://en.wikipedia.org/wiki/Decisiontreelearning#Gini_impurity

BetaML.Utils.generateParallelRngs — Method

generateParallelRngs(rng::AbstractRNG, n::Integer;reSeed=false)

For multi-threaded models, return n independent random number generators (one per thread) to be used in threaded computations.

Note that each ring is a copy of the original random ring. This means that code that use these RNGs will not change the original RNG state.

Use it with rngs = generateParallelRngs(rng,Threads.nthreads()) to have a separate rng per thread. By default the function doesn't re-seed the RNG, as you may want to have a loop index based re-seeding strategy rather than a threadid-based one (to guarantee the same result independently of the number of threads). If you prefer, you can instead re-seed the RNG here (using the parameter reSeed=true), such that each thread has a different seed. Be aware however that the stream of number generated will depend from the number of threads at run time.

BetaML.Utils.getPermutations — Method

getPermutations(v::AbstractArray{T,1};keepStructure=false)

Return a vector of either (a) all possible permutations (uncollected) or (b) just those based on the unique values of the vector

Useful to measure accuracy where you don't care about the actual name of the labels, like in unsupervised classifications (e.g. clustering)

BetaML.Utils.getScaleFactors — Method

getScaleFactors(x;skip)

Return the scale factors (for each dimensions) in order to scale a matrix X (n,d) such that each dimension has mean 0 and variance 1.

Parameters

x: the (n × d) dimension matrix to scale on each dimension d

skip: an array of dimension index to skip the scaling [def: []]

Return

A touple whose first elmement is the shift and the second the multiplicative

term to make the scale.

BetaML.Utils.gini — Method

gini(x)

Calculate the Gini Impurity for a list of items (or rows).

See: https://en.wikipedia.org/wiki/Decisiontreelearning#Information_gain

BetaML.Utils.integerDecoder — Method

integerDecoder(x,factors::AbstractVector{T};unique)

Decode an array of integers to an array of T corresponding to the elements of factors

Parameters:

x: The vector to decode

factors: The vector of elements to use for the encoding

unique: Wether factors is already made of unique elements [def: true]

Return:

A vector of length(x) elements corresponding to the (unique) factors elements at the position x

Example:

julia> integerDecoder([1, 2, 2, 3, 2, 1],["aa","cc","bb"]) # out: ["aa","cc","cc","bb","cc","aa"]

BetaML.Utils.integerEncoder — Method

integerEncoder(x;factors=unique(x))

Encode an array of T to an array of integers using the their position in factor vector (default to the unique vector of the input array)

Parameters:

x: The vector to encode

factors: The vector of factors whose position is the result of the encoding [def: unique(x)]

Return:

A vector of [1,length(x)] integers corresponding to the position of each element in the factors vector`

Note:

Attention that while this function creates a ordered (and sortable) set, it is up to the user to be sure that this "property" is not indeed used in his code if the unencoded data is indeed unordered.

Example:

julia> integerEncoder(["a","e","b","e"],factors=["a","b","c","d","e"]) # out: [1,5,2,5]

BetaML.Utils.issortable — Method

Return wheather an array is sortable, i.e. has methos issort defined

BetaML.Utils.l1_distance — Method

L1 norm distance (aka Manhattan Distance)

BetaML.Utils.l2_distance — Method

Euclidean (L2) distance

BetaML.Utils.l2²_distance — Method

Squared Euclidean (L2) distance

BetaML.Utils.lse — Method

LogSumExp for efficiently computing log(sum(exp.(x)))

BetaML.Utils.makeMatrix — Method

Transform an Array{T,1} in an Array{T,2} and leave unchanged Array{T,2}.

BetaML.Utils.meanDicts — Method

meanDicts(dicts)

Compute the mean of the values of an array of dictionaries.

Given dicts an array of dictionaries, meanDicts first compute the union of the keys and then average the values. If the original valueas are probabilities (non-negative items summing to 1), the result is also a probability distribution.

BetaML.Utils.meanRelError — Method

meanRelError(ŷ,y;normDim=true,normRec=true,p=1)

Compute the mean relative error (l-1 based by default) between ŷ and y.

There are many ways to compute a mean relative error. In particular, if normRec (normDim) is set to true, the records (dimensions) are normalised, in the sense that it doesn't matter if a record (dimension) is bigger or smaller than the others, the relative error is first computed for each record (dimension) and then it is averaged. With both normDim and normRec set to false the function returns the relative mean error; with both set to true (default) it returns the mean relative error (i.e. with p=1 the "mean absolute percentage error (MAPE)") The parameter p [def: 1] controls the p-norm used to define the error.

The mean relative error enfatises the relativeness of the error, i.e. all observations and dimensions weigth the same, wether large or small. Conversly, in the relative mean error the same relative error on larger observations (or dimensions) weights more.

For example, given y = [1,44,3] and ŷ = [2,45,2], the mean relative error meanRelError(ŷ,y) is 0.452, while the relative mean error meanRelError(ŷ,y, normRec=false) is "only" 0.0625.

BetaML.Utils.mish — Method

mish(x)

https://arxiv.org/pdf/1908.08681v1.pdf

BetaML.Utils.mode — Method

mode(dicts)

Given a vector of dictionaries representing probabilities it returns the mode of each element in terms of the key

Use it to return a unique value from a multiclass classifier returning probabilities.

BetaML.Utils.oneHotEncoder — Method

oneHotEncoder(x;d,factors,count)

Encode arrays (or arrays of arrays) of categorical data as matrices of one column per factor.

The case of arrays of arrays is for when at each record you have more than one categorical output. You can then decide to encode just the presence of the factors or their counting

Parameters:

x: The data to convert (array or array of arrays)

d: The number of dimensions in the output matrix [def: maximum(x) for integers and length(factors) otherwise]

factors: The factors from which to encode [def: 1:d for integer x or unique(x) otherwise]

count: Wether to count multiple instances on the same dimension/record (true) or indicate just presence. [def: false]

Examples

julia> oneHotEncoder(["a","c","c"],factors=["a","b","c","d"])
3×4 Matrix{Int64}:
 1  0  0  0
 0  0  1  0
 0  0  1  0
julia> oneHotEncoder([2,4,4])
3×4 Matrix{Int64}:
 0  1  0  0
 0  0  0  1
 0  0  0  1
 julia> oneHotEncoder([[2,2,1],[2,4,4]],count=true)
2×4 Matrix{Int64}:
 1  2  0  0
 0  1  0  2

BetaML.Utils.pca — Method

pca(X;K,error)

Perform Principal Component Analysis returning the matrix reprojected among the dimensions of maximum variance.

Parameters:

X : The (N,D) data to reproject

K : The number of dimensions to maintain (with K<=D) [def: nothing]

error: The maximum approximation error that we are willing to accept [def: 0.05]

Return:

A named tuple with:

X: The reprojected (NxK) matrix with the column dimensions organized in descending order of of the proportion of explained variance

K: The number of dimensions retieved

error: The actual proportion of variance not explained in the reprojected dimensions

P: The (D,K) matrix of the eigenvectors associated to the K-largest eigenvalues used to reproject the data matrix

explVarByDim: An array of dimensions D with the share of the cumulative variance explained by dimensions (the last element being always 1.0)

Notes:

If K is provided, the parameter error has no effect.

If one doesn't know a priori the error that she/he is willling to accept, nor the wished number of dimensions, he/she can run this pca function with out = pca(X,K=size(X,2)) (i.e. with K=D), analise the proportions of explained cumulative variance by dimensions in out.explVarByDim, choose the number of dimensions K according to his/her needs and finally pick from the reprojected matrix only the number of dimensions needed, i.e. out.X[:,1:K].

Example:

julia> X = [1 10 100; 1.1 15 120; 0.95 23 90; 0.99 17 120; 1.05 8 90; 1.1 12 95]
6×3 Array{Float64,2}:
 1.0   10.0  100.0
 1.1   15.0  120.0
 0.95  23.0   90.0
 0.99  17.0  120.0
 1.05   8.0   90.0
 1.1   12.0   95.0
 julia> X = pca(X,error=0.05).X
6×2 Array{Float64,2}:
  3.1783   100.449
  6.80764  120.743
 16.8275    91.3551
  8.80372  120.878
  1.86179   90.3363
  5.51254   95.5965

BetaML.Utils.plu — Method

plu(x;α=0.1,c=1)

Piecewise Linear Unit

https://arxiv.org/pdf/1809.09534.pdf

BetaML.Utils.polynomialKernel — Method

Polynomial kernel parametrised with c=0 and d=2 (i.e. a quadratic kernel). For other cᵢ and dᵢ use K = (x,y) -> polynomialKernel(x,y,c=cᵢ,d=dᵢ) as kernel function in the supporting algorithms

BetaML.Utils.radialKernel — Method

Radial Kernel (aka RBF kernel) parametrised with γ=1/2. For other gammas γᵢ use K = (x,y) -> radialKernel(x,y,γ=γᵢ) as kernel function in the supporting algorithms

BetaML.Utils.relu — Method

relu(x)

Rectified Linear Unit

https://www.cs.toronto.edu/~hinton/absps/reluICML.pdf

BetaML.Utils.scale — Function

scale(x,scaleFactors;rev)

Perform a linear scaling of x using scaling factors scaleFactors.

Parameters

x: The (n × d) dimension matrix to scale on each dimension d

scalingFactors: A tuple of the constant and multiplicative scaling factor

respectively [def: the scaling factors needed to scale x to mean 0 and variance 1]

rev: Whether to invert the scaling [def: false]

Return

The scaled matrix

Notes:

Also available scale!(x,scaleFactors) for in-place scaling.

Retrieve the scale factors with the getScaleFactors() function

BetaML.Utils.sigmoid — Method

sigmoid(x)

BetaML.Utils.singleUnique — Method

singleUnique(x) Return the unique values of x whether x is an array of arrays, an array or a scalar

BetaML.Utils.softmax — Method

softmax (x; β=1)

The input x is a vector. Return a PMF

BetaML.Utils.softplus — Method

softplus(x)

https://en.wikipedia.org/wiki/Rectifier(neuralnetworks)#Softplus

BetaML.Utils.squaredCost — Method

squaredCost(ŷ,y)

Compute the squared costs between a vector of prediction and one of observations as (1/2)*norm(y - ŷ)^2.

Aside the 1/2 term, it correspond to the squared l-2 norm distance and when it is averaged on multiple datapoints corresponds to the Mean Squared Error (MSE). It is mostly used for regression problems.

BetaML.Utils.sterling — Method

Sterling number: number of partitions of a set of n elements in k sets

BetaML.Utils.variance — Method

variance(x) - population variance