The BetaML.Clustering Module

BetaML.Clustering — Module

Clustering module (WIP)

(Hard) Clustering algorithms

Provide hard clustering methods using K-means and k-medoids. Please see also the GMM module for GMM-based soft clustering (i.e. where a probability distribution to be part of the various classes is assigned to each record instead of a single class), missing values imputation / collaborative filtering / reccomendation systems using clustering methods as backend.

The module provides the following models. Use ?[model] to access their documentation:

KMeansClusterer: Classical KMean algorithm

KMedoidsClusterer: Kmedoids algorithm with configurable distance metric

Module Index

BetaML.Clustering.KMeans

BetaML.Clustering.KMeansClusterer

BetaML.Clustering.KMeansMedoidsHyperParametersSet

BetaML.Clustering.KMedoids

BetaML.Clustering.KMedoidsClusterer

BetaML.Clustering.kmeans

BetaML.Clustering.kmedoids

Detailed API

BetaML.Clustering.KMeans — Type

mutable struct KMeans <: MLJModelInterface.Unsupervised

The classical KMeans clustering algorithm, from the Beta Machine Learning Toolkit (BetaML).

Parameters:

n_classes::Int64

Number of classes to discriminate the data [def: 3]

dist::Function

Function to employ as distance. Default to the Euclidean distance. Can be one of the predefined distances (l1_distance, l2_distance, l2squared_distance), cosine_distance), any user defined function accepting two vectors and returning a scalar or an anonymous function with the same characteristics. Attention that, contrary to KMedoids, the KMeansClusterer algorithm is not guaranteed to converge with other distances than the Euclidean one.

initialisation_strategy::String

The computation method of the vector of the initial representatives. One of the following:

"random": randomly in the X space

"grid": using a grid approach

"shuffle": selecting randomly within the available points [default]

"given": using a provided set of initial representatives provided in the initial_representatives parameter

initial_representatives::Union{Nothing, Matrix{Float64}}

Provided (K x D) matrix of initial representatives (useful only with initialisation_strategy="given") [default: nothing]

rng::Random.AbstractRNG

Random Number Generator [deafult: Random.GLOBAL_RNG]

Notes:

data must be numerical

online fitting (re-fitting with new data) is supported

BetaML.Clustering.KMeansClusterer — Type

mutable struct KMeansClusterer <: BetaMLUnsupervisedModel

The classical "K-Means" clustering algorithm (unsupervised).

For the parameters see ?KMeansMedoidsHyperParametersSet and ?BetaMLDefaultOptionsSet.

Notes:

data must be numerical

online fitting (re-fitting with new data) is supported

BetaML.Clustering.KMeansMedoidsHyperParametersSet — Type

mutable struct KMeansMedoidsHyperParametersSet <: BetaMLHyperParametersSet

Hyperparameters for both the KMeansClusterer and KMedoidsClusterer models

Parameters:

n_classes::Int64

Number of classes to discriminate the data [def: 3]

dist::Function

Function to employ as distance. Default to the Euclidean distance. Can be one of the predefined distances (l1_distance, l2_distance, l2squared_distance), cosine_distance), any user defined function accepting two vectors and returning a scalar or an anonymous function with the same characteristics. Attention that the KMeansClusterer algorithm is not guaranteed to converge with other distances than the Euclidean one.

initialisation_strategy::String

The computation method of the vector of the initial representatives. One of the following:

"random": randomly in the X space [default]

"grid": using a grid approach

"shuffle": selecting randomly within the available points

"given": using a provided set of initial representatives provided in the initial_representatives parameter

initial_representatives::Union{Nothing, Matrix{Float64}}

Provided (K x D) matrix of initial representatives (useful only with initialisation_strategy="given") [default: nothing]

BetaML.Clustering.KMedoids — Type

mutable struct KMedoids <: MLJModelInterface.Unsupervised

Parameters:

n_classes::Int64

Number of classes to discriminate the data [def: 3]

dist::Function

Function to employ as distance. Default to the Euclidean distance. Can be one of the predefined distances (l1_distance, l2_distance, l2squared_distance), cosine_distance), any user defined function accepting two vectors and returning a scalar or an anonymous function with the same characteristics.

initialisation_strategy::String

The computation method of the vector of the initial representatives. One of the following:

"random": randomly in the X space

"grid": using a grid approach

"shuffle": selecting randomly within the available points [default]

"given": using a provided set of initial representatives provided in the initial_representatives parameter

initial_representatives::Union{Nothing, Matrix{Float64}}

Provided (K x D) matrix of initial representatives (useful only with initialisation_strategy="given") [default: nothing]

rng::Random.AbstractRNG

Random Number Generator [deafult: Random.GLOBAL_RNG]

The K-medoids clustering algorithm with customisable distance function, from the Beta Machine Learning Toolkit (BetaML).

Similar to K-Means, but the "representatives" (the cetroids) are guaranteed to be one of the training points. The algorithm work with any arbitrary distance measure.

Notes:

data must be numerical

online fitting (re-fitting with new data) is supported

BetaML.Clustering.KMedoidsClusterer — Type

mutable struct KMedoidsClusterer <: BetaMLUnsupervisedModel

The classical "K-Medoids" clustering algorithm (unsupervised).

Similar to K-Means, but the "representatives" (the cetroids) are guaranteed to be one of the training points. The algorithm work with any arbitrary distance measure.

For the parameters see ?KMeansMedoidsHyperParametersSet and ?BetaMLDefaultOptionsSet.

Notes:

data must be numerical

online fitting (re-fitting with new data) is supported

BetaML.Clustering.kmeans — Method

kmeans(X,K;dist,initialisationstrategy,initialrepresentatives)

Compute K-Mean algorithm to identify K clusters of X using Euclidean distance

Warning

This function is deprecated and will possibly be removed in BetaML 0.9. Use KMeansClusterer instead.

Parameters:

X: a (N x D) data to clusterise

K: Number of cluster wonted

dist: Function to employ as distance (see notes). Default to Euclidean distance.

initialisation_strategy: Whether to select the initial representative vectors:

random: randomly in the X space

grid: using a grid approach [default]

shuffle: selecting randomly within the available points

given: using a provided set of initial representatives provided in the initial_representatives parameter

initial_representatives: Provided (K x D) matrix of initial representatives (used only together with the given initialisation_strategy) [default: nothing]

rng: Random Number Generator (see FIXEDSEED) [deafult: Random.GLOBAL_RNG]

Returns:

A tuple of two items, the first one being a vector of size N of ids of the clusters associated to each point and the second one the (K x D) matrix of representatives

Notes:

Some returned clusters could be empty

The dist parameter can be:

Any user defined function accepting two vectors and returning a scalar

An anonymous function with the same characteristics (e.g. dist = (x,y) -> norm(x-y)^2)

One of the above predefined distances: l1_distance, l2_distance, l2squared_distance, cosine_distance

Example:

julia> (clIdx,Z) = kmeans([1 10.5;1.5 10.8; 1.8 8; 1.7 15; 3.2 40; 3.6 32; 3.3 38; 5.1 -2.3; 5.2 -2.4],3)

BetaML.Clustering.kmedoids — Method

kmedoids(X,K;dist,initialisationstrategy,initialrepresentatives)

Compute K-Medoids algorithm to identify K clusters of X using distance definition dist

Warning

This function is deprecated and will possibly be removed in BetaML 0.9. Use KMedoidsClusterer instead.

Parameters:

X: a (n x d) data to clusterise

K: Number of cluster wonted

dist: Function to employ as distance (see notes). Default to Euclidean distance.

initialisation_strategy: Whether to select the initial representative vectors:

random: randomly in the X space

grid: using a grid approach

shuffle: selecting randomly within the available points [default]

given: using a provided set of initial representatives provided in the initial_representatives parameter

initial_representatives: Provided (K x D) matrix of initial representatives (used only together with the given initialisation_strategy) [default: nothing]

rng: Random Number Generator (see FIXEDSEED) [deafult: Random.GLOBAL_RNG]

Returns:

A tuple of two items, the first one being a vector of size N of ids of the clusters associated to each point and the second one the (K x D) matrix of representatives

Notes:

Some returned clusters could be empty

The dist parameter can be:

Any user defined function accepting two vectors and returning a scalar

An anonymous function with the same characteristics (e.g. dist = (x,y) -> norm(x-y)^2)

One of the above predefined distances: l1_distance, l2_distance, l2squared_distance, cosine_distance

Example:

julia> (clIdx,Z) = kmedoids([1 10.5;1.5 10.8; 1.8 8; 1.7 15; 3.2 40; 3.6 32; 3.3 38; 5.1 -2.3; 5.2 -2.4],3,initialisation_strategy="grid")