The BetaML.Clustering Module

BetaML.Clustering — Module

Clustering module (WIP)

(Hard) Clustering algorithms

Provide hard clustering methods using K-means and k-medoids. Please see also the GMM module for GMM-based soft clustering (i.e. where a probability distribution to be part of the various classes is assigned to each record instead of a single class), missing values imputation / collaborative filtering / reccomendation systems using clustering methods as backend.

The module provides the following models. Use ?[model] to access their documentation:

KMeansClusterer: Classical KMean algorithm

KMedoidsClusterer: Kmedoids algorithm with configurable distance metric

Module Index

BetaML.Clustering.KMeans

BetaML.Clustering.KMeansClusterer

BetaML.Clustering.KMeansMedoidsHyperParametersSet

BetaML.Clustering.KMedoids

BetaML.Clustering.KMedoidsClusterer

Detailed API

BetaML.Clustering.KMeans — Type

mutable struct KMeans <: MLJModelInterface.Unsupervised

The classical KMeans clustering algorithm, from the Beta Machine Learning Toolkit (BetaML).

Parameters:

n_classes::Int64

: Number of classes to discriminate the data [def: 3]

dist::Function

: Function to employ as distance. Default to the Euclidean distance. Can be one of the predefined distances (l1_distance, l2_distance, l2squared_distance), cosine_distance), any user defined function accepting two vectors and returning a scalar or an anonymous function with the same characteristics. Attention that, contrary to KMedoids, the KMeansClusterer algorithm is not guaranteed to converge with other distances than the Euclidean one.

initialisation_strategy::String

: The computation method of the vector of the initial representatives. One of the following: - "random": randomly in the X space - "grid": using a grid approach - "shuffle": selecting randomly within the available points [default] - "given": using a provided set of initial representatives provided in the initial_representatives parameter

initial_representatives::Union{Nothing, Matrix{Float64}}

: Provided (K x D) matrix of initial representatives (useful only with initialisation_strategy="given") [default: nothing]

rng::Random.AbstractRNG

: Random Number Generator [deafult: Random.GLOBAL_RNG]

Notes:

data must be numerical

online fitting (re-fitting with new data) is supported

Example:

julia> using MLJ

julia> modelType                   = @load KMeans pkg = "BetaML"
[ Info: For silent loading, specify `verbosity=0`. 
import BetaML ✔
BetaML.Clustering.KMeans

julia> model                       = modelType()
KMeans(
 n_classes = 3, 
 dist = BetaML.Clustering.var"#25#27"(), 
 initialisation_strategy = "shuffle", 
 initial_representatives = nothing, 
 rng = Random._GLOBAL_RNG())

julia> X, y                        = @load_iris;

julia> (fitResults, cache, report) = MLJ.fit(model, 0, X);

julia> est_classes                 = predict(model, fitResults, X)
150-element CategoricalArrays.CategoricalArray{Int64,1,UInt32}:
3
3
3
⋮
1
1
2

BetaML.Clustering.KMeansClusterer — Type

mutable struct KMeansClusterer <: BetaMLUnsupervisedModel

The classical "K-Means" clustering algorithm (unsupervised).

Learn to partition the data and assign each record to one of the n_classes classes according to a distance metric (default Euclidean).

For the parameters see ?KMeansMedoidsHyperParametersSet and ?BetaMLDefaultOptionsSet.

Notes:

data must be numerical

online fitting (re-fitting with new data) is supported

Example :

julia> using BetaML

julia> X = [1.1 10.1; 0.9 9.8; 10.0 1.1; 12.1 0.8; 0.8 9.8]
5×2 Matrix{Float64}:
  1.1  10.1
  0.9   9.8
 10.0   1.1
 12.1   0.8
  0.8   9.8

julia> mod = KMeansClusterer(n_classes=2)
KMeansClusterer - A K-Means Model (unfitted)

julia> classes = fit!(mod,X)
5-element Vector{Int64}:
 1
 1
 2
 2
 1

julia> newclasses = fit!(mod,[11 0.9])
1-element Vector{Int64}:
 2

julia> info(mod)
Dict{String, Any} with 2 entries:
  "fitted_records" => 6
  "xndims"         => 2

julia> parameters(mod)
BetaML.Clustering.KMeansMedoidsLearnableParameters (a BetaMLLearnableParametersSet struct)
- representatives: [1.13366 9.7209; 11.0 0.9]

BetaML.Clustering.KMeansMedoidsHyperParametersSet — Type

mutable struct KMeansMedoidsHyperParametersSet <: BetaMLHyperParametersSet

Hyperparameters for both the KMeansClusterer and KMedoidsClusterer models

Parameters:

n_classes::Int64

: Number of classes to discriminate the data [def: 3]

dist::Function

: Function to employ as distance. Default to the Euclidean distance. Can be one of the predefined distances (l1_distance, l2_distance, l2squared_distance), cosine_distance), any user defined function accepting two vectors and returning a scalar or an anonymous function with the same characteristics. Attention that the KMeansClusterer algorithm is not guaranteed to converge with other distances than the Euclidean one.

initialisation_strategy::String

: The computation method of the vector of the initial representatives. One of the following: - "random": randomly in the X space [default] - "grid": using a grid approach - "shuffle": selecting randomly within the available points - "given": using a provided set of initial representatives provided in the initial_representatives parameter

initial_representatives::Union{Nothing, Matrix{Float64}}

: Provided (K x D) matrix of initial representatives (useful only with initialisation_strategy="given") [default: nothing]

BetaML.Clustering.KMedoids — Type

mutable struct KMedoids <: MLJModelInterface.Unsupervised

Parameters:

n_classes::Int64

: Number of classes to discriminate the data [def: 3]

dist::Function

: Function to employ as distance. Default to the Euclidean distance. Can be one of the predefined distances (l1_distance, l2_distance, l2squared_distance), cosine_distance), any user defined function accepting two vectors and returning a scalar or an anonymous function with the same characteristics.

initialisation_strategy::String

: The computation method of the vector of the initial representatives. One of the following: - "random": randomly in the X space - "grid": using a grid approach - "shuffle": selecting randomly within the available points [default] - "given": using a provided set of initial representatives provided in the initial_representatives parameter

initial_representatives::Union{Nothing, Matrix{Float64}}

: Provided (K x D) matrix of initial representatives (useful only with initialisation_strategy="given") [default: nothing]

rng::Random.AbstractRNG

: Random Number Generator [deafult: Random.GLOBAL_RNG]

The K-medoids clustering algorithm with customisable distance function, from the Beta Machine Learning Toolkit (BetaML).

Similar to K-Means, but the "representatives" (the cetroids) are guaranteed to be one of the training points. The algorithm work with any arbitrary distance measure.

Notes:

data must be numerical

online fitting (re-fitting with new data) is supported

Example:

julia> using MLJ

julia> modelType                   = @load KMedoids pkg = "BetaML"
[ Info: For silent loading, specify `verbosity=0`. 
import BetaML ✔
BetaML.Clustering.KMedoids

julia> model                       = modelType()
KMedoids(
  n_classes = 3, 
  dist = BetaML.Clustering.var"#49#51"(), 
  initialisation_strategy = "shuffle", 
  initial_representatives = nothing, 
  rng = Random._GLOBAL_RNG())

julia> X, y                        = @load_iris;

julia> (fitResults, cache, report) = MLJ.fit(model, 0, X);

julia> est_classes                 = predict(model, fitResults, X)
150-element CategoricalArrays.CategoricalArray{Int64,1,UInt32}:
 2
 3
 3
 ⋮
 1
 1
 1

BetaML.Clustering.KMedoidsClusterer — Type

mutable struct KMedoidsClusterer <: BetaMLUnsupervisedModel

The classical "K-Medoids" clustering algorithm (unsupervised).

Similar to K-Means, learn to partition the data and assign each record to one of the n_classes classes according to a distance metric, but the "representatives" (the cetroids) are guaranteed to be one of the training points. The algorithm work with any arbitrary distance measure (default Euclidean).

For the parameters see ?KMeansMedoidsHyperParametersSet and ?BetaMLDefaultOptionsSet.

Notes:

data must be numerical

online fitting (re-fitting with new data) is supported

Example:

julia> using BetaML

julia> X = [1.1 10.1; 0.9 9.8; 10.0 1.1; 12.1 0.8; 0.8 9.8] 5×2 Matrix{Float64}: 1.1 10.1 0.9 9.8 10.0 1.1 12.1 0.8 0.8 9.8

julia> mod = KMedoidsClusterer(n_classes=2) KMedoidsClusterer - A K-Medoids Model (unfitted)

julia> classes = fit!(mod,X) 5-element Vector{Int64}: 1 1 2 2 1

julia> newclasses = fit!(mod,[11 0.9]) 1-element Vector{Int64}: 2

julia> info(mod) Dict{String, Any} with 2 entries: "fitted_records" => 6 "xndims" => 2

julia> parameters(mod) BetaML.Clustering.KMeansMedoidsLearnableParameters (a BetaMLLearnableParametersSet struct)

representatives: [0.9 9.8; 11.0 0.9]