The BetaML.GMM Module

BetaML.GMM — Module

GMM module

Generative (Gaussian) Mixed Model learners (supervised/unsupervised)

Provides clustering and regressors using (Generative) Gaussiam Mixture Model (probabilistic).

Collaborative filtering / missing values imputation / reccomendation systems based on GMM is available in the Imputation module.

The module provides the following models. Use ?[model] to access their documentation:

GMMClusterer: soft-clustering using GMM

GMMRegressor1: regressor using GMM as back-end (first algorithm)

GMMRegressor1: regressor using GMM as back-end (second algorithm)

All the algorithms works with arbitrary mixture distribution, altought only {Spherical|Diagonal|Full} Gaussian mixtures has been implemented. User defined mixtures can be used defining a struct as subtype of AbstractMixture and implementing for that mixture the following functions:

init_mixtures!(mixtures, X; minimum_variance, minimum_covariance, initialisation_strategy)

lpdf(m,x,mask) (for the e-step)

update_parameters!(mixtures, X, pₙₖ; minimum_variance, minimum_covariance) (the m-step)

All the GMM-based algorithms works only with numerical data, but accepts also Missing one.

Module Index

BetaML.GMM.GMMClusterer

BetaML.GMM.GMMHyperParametersSet

BetaML.GMM.GMMRegressor1

BetaML.GMM.GMMRegressor2

BetaML.GMM.GaussianMixtureClusterer

BetaML.GMM.GaussianMixtureRegressor

BetaML.GMM.MultitargetGaussianMixtureRegressor

BetaML.GMM.gmm

BetaML.GMM.init_mixtures!

BetaML.GMM.lpdf

BetaML.GMM.lpdf

BetaML.GMM.lpdf

Detailed API

BetaML.GMM.GMMClusterer — Type

mutable struct GMMClusterer <: BetaMLUnsupervisedModel

Assign class probabilities to records (i.e. soft clustering) assuming a probabilistic generative model of observed data using mixtures.

For the parameters see ?GMMHyperParametersSet and ?BetaMLDefaultOptionsSet.

Notes:

Data must be numerical

Mixtures can be user defined: see the ?GMM module documentation for a discussion on provided vs custom mixtures.

Online fitting (re-fitting with new data) is supported by setting the old learned mixtrures as the starting values

The model is fitted using an Expectation-Minimisation (EM) algorithm that supports Missing data and is implemented in the log-domain for better numerical accuracy with many dimensions

BetaML.GMM.GMMHyperParametersSet — Type

mutable struct GMMHyperParametersSet <: BetaMLHyperParametersSet

Hyperparameters for GMM clusters and other GMM-based algorithms

Parameters:

n_classes

Number of mixtures (latent classes) to consider [def: 3]

initial_probmixtures

Initial probabilities of the categorical distribution (n_classes x 1) [default: []]

mixtures

An array (of length n_classes) of the mixtures to employ (see the [?GMM](@ref GMM) module). Each mixture object can be provided with or without its parameters (e.g. mean and variance for the gaussian ones). Fully qualified mixtures are useful only if theinitialisationstrategyparameter is set to "gived" [def: `[DiagonalGaussian() for i in 1:nclasses]`]

tol

Tolerance to stop the algorithm [default: 10^(-6)]

minimum_variance

Minimum variance for the mixtures [default: 0.05]

minimum_covariance

Minimum covariance for the mixtures with full covariance matrix [default: 0]. This should be set different than minimum_variance (see notes).

initialisation_strategy

The computation method of the vector of the initial mixtures. One of the following:

"grid": using a grid approach

"given": using the mixture provided in the fully qualified mixtures parameter

"kmeans": use first kmeans (itself initialised with a "grid" strategy) to set the initial mixture centers [default]

Note that currently "random" and "shuffle" initialisations are not supported in gmm-based algorithms.

maximum_iterations

Maximum number of iterations [def: typemax(Int64), i.e. ∞]

BetaML.GMM.GMMRegressor1 — Type

mutable struct GMMRegressor1 <: BetaMLUnsupervisedModel

A multi-dimensional, missing data friendly non-linear regressor based on Generative (Gaussian) Mixture Model (strategy "1").

The training data is used to fit a probabilistic model with latent mixtures (Gaussian distributions with different covariances are already implemented) and then predictions of new data is obtained by fitting the new data to the mixtures.

For hyperparameters see GMMHyperParametersSet and BetaMLDefaultOptionsSet.

This strategy (GMMRegressor1) works by fitting the EM algorithm on the feature matrix X. Once the data has been probabilistically assigned to the various classes, a mean value of fitting values Y is computed for each cluster (using the probabilities as weigths). At predict time, the new data is first fitted to the learned mixtures using the e-step part of the EM algorithm to obtain the probabilistic assignment of each record to the various mixtures. Then these probabilities are multiplied to the mixture averages for the Y dimensions learned at training time to obtain the predicted value(s) for each record.

Notes:

the predicted values are always a matrix, even when a single variable is predicted (use dropdims(ŷ,dims=2) to get a single vector).

BetaML.GMM.GMMRegressor2 — Type

mutable struct GMMRegressor2 <: BetaMLUnsupervisedModel

A multi-dimensional, missing data friendly non-linear regressor based on Generative (Gaussian) Mixture Model.

The training data is used to fit a probabilistic model with latent mixtures (Gaussian distributions with different covariances are already implemented) and then predictions of new data is obtained by fitting the new data to the mixtures.

For hyperparameters see GMMHyperParametersSet and BetaMLDefaultOptionsSet.

Thsi strategy (GMMRegressor2) works by training the EM algorithm on a combined (hcat) matrix of X and Y. At predict time, the new data is first fitted to the learned mixtures using the e-step part of the EM algorithm (and using missing values for the dimensions belonging to Y) to obtain the probabilistic assignment of each record to the various mixtures. Then these probabilities are multiplied to the mixture averages for the Y dimensions to obtain the predicted value(s) for each record.

BetaML.GMM.GaussianMixtureClusterer — Type

mutable struct GaussianMixtureClusterer <: MLJModelInterface.Unsupervised

A Expectation-Maximisation clustering algorithm with customisable mixtures, from the Beta Machine Learning Toolkit (BetaML).

Hyperparameters:

n_classes::Int64

Number of mixtures (latent classes) to consider [def: 3]

initial_probmixtures::AbstractVector{Float64}

Initial probabilities of the categorical distribution (n_classes x 1) [default: []]

mixtures::Vector{AbstractMixture}

An array (of length n_classes) of the mixtures to employ (see the [?GMM](@ref GMM) module of BetaML for available mixtures). Each mixture object can be provided with or without its parameters (e.g. mean and variance for the gaussian ones). Fully qualified mixtures are useful only if theinitialisationstrategyparameter is set to "given" [def: `[DiagonalGaussian() for i in 1:nclasses]`]

tol::Float64

Tolerance to stop the algorithm [default: 10^(-6)]

minimum_variance::Float64

Minimum variance for the mixtures [default: 0.05]

minimum_covariance::Float64

Minimum covariance for the mixtures with full covariance matrix [default: 0]. This should be set different than minimum_variance (see notes).

initialisation_strategy::String

The computation method of the vector of the initial mixtures. One of the following:

"grid": using a grid approach

"given": using the mixture provided in the fully qualified mixtures parameter

"kmeans": use first kmeans (itself initialised with a "grid" strategy) to set the initial mixture centers [default]

Note that currently "random" and "shuffle" initialisations are not supported in gmm-based algorithms.

maximum_iterations::Int64

Maximum number of iterations [def: typemax(Int64), i.e. ∞]

rng::Random.AbstractRNG

Random Number Generator [deafult: Random.GLOBAL_RNG]

BetaML.GMM.GaussianMixtureRegressor — Type

mutable struct GaussianMixtureRegressor <: MLJModelInterface.Deterministic

A non-linear regressor derived from fitting the data on a probabilistic model (Gaussian Mixture Model). Relatively fast.

This is the single-target version of the model. If you want to predict several labels (y) at once, use the MLJ model MultitargetGaussianMixtureRegressor.

Hyperparameters:

n_classes::Int64

Number of mixtures (latent classes) to consider [def: 3]

initial_probmixtures::Vector{Float64}

Initial probabilities of the categorical distribution (n_classes x 1) [default: []]

mixtures::Vector{AbstractMixture}

An array (of length n_classes) of the mixtures to employ (see the [?GMM](@ref GMM) module). Each mixture object can be provided with or without its parameters (e.g. mean and variance for the gaussian ones). Fully qualified mixtures are useful only if theinitialisationstrategyparameter is set to "given" [def: `[DiagonalGaussian() for i in 1:nclasses]`]

tol::Float64

Tolerance to stop the algorithm [default: 10^(-6)]

minimum_variance::Float64

Minimum variance for the mixtures [default: 0.05]

minimum_covariance::Float64

Minimum covariance for the mixtures with full covariance matrix [default: 0]. This should be set different than minimum_variance (see notes).

initialisation_strategy::String

The computation method of the vector of the initial mixtures. One of the following:

"grid": using a grid approach

"given": using the mixture provided in the fully qualified mixtures parameter

"kmeans": use first kmeans (itself initialised with a "grid" strategy) to set the initial mixture centers [default]

Note that currently "random" and "shuffle" initialisations are not supported in gmm-based algorithms.

maximum_iterations::Int64

Maximum number of iterations [def: typemax(Int64), i.e. ∞]

rng::Random.AbstractRNG

Random Number Generator [deafult: Random.GLOBAL_RNG]

BetaML.GMM.MultitargetGaussianMixtureRegressor — Type

mutable struct MultitargetGaussianMixtureRegressor <: MLJModelInterface.Deterministic

A non-linear regressor derived from fitting the data on a probabilistic model (Gaussian Mixture Model). Relatively fast.

This is the multi-target version of the model. If you want to predict a single label (y), use the MLJ model GaussianMixtureRegressor.

Hyperparameters:

n_classes::Int64

Number of mixtures (latent classes) to consider [def: 3]

initial_probmixtures::Vector{Float64}

Initial probabilities of the categorical distribution (n_classes x 1) [default: []]

mixtures::Vector{AbstractMixture}

An array (of length n_classes) of the mixtures to employ (see the [?GMM](@ref GMM) module). Each mixture object can be provided with or without its parameters (e.g. mean and variance for the gaussian ones). Fully qualified mixtures are useful only if theinitialisationstrategyparameter is set to "given" [def: `[DiagonalGaussian() for i in 1:nclasses]`]

tol::Float64

Tolerance to stop the algorithm [default: 10^(-6)]

minimum_variance::Float64

Minimum variance for the mixtures [default: 0.05]

minimum_covariance::Float64

Minimum covariance for the mixtures with full covariance matrix [default: 0]. This should be set different than minimum_variance (see notes).

initialisation_strategy::String

The computation method of the vector of the initial mixtures. One of the following:

"grid": using a grid approach

"given": using the mixture provided in the fully qualified mixtures parameter

"kmeans": use first kmeans (itself initialised with a "grid" strategy) to set the initial mixture centers [default]

Note that currently "random" and "shuffle" initialisations are not supported in gmm-based algorithms.

maximum_iterations::Int64

Maximum number of iterations [def: typemax(Int64), i.e. ∞]

rng::Random.AbstractRNG

Random Number Generator [deafult: Random.GLOBAL_RNG]

BetaML.GMM.gmm — Method

gmm(X,K;initialprobmixtures,mixtures,tol,verbosity,minimumvariance,minimumcovariance,initialisationstrategy)

Compute Expectation-Maximisation algorithm to identify K clusters of X data, i.e. employ a Generative Mixture Model as the underlying probabilistic model.

Warning

This function is deprecated and will possibly be removed in BetaML 0.9. Use one of the various models that use GMM as backend instead.

X can contain missing values in some or all of its dimensions. In such case the learning is done only with the available data. Implemented in the log-domain for better numerical accuracy with many dimensions.

Parameters:

X : A (n x d) data to clusterise

K : Number of cluster wanted

initial_probmixtures : Initial probabilities of the categorical distribution (K x 1) [default: []]

mixtures: An array (of length K) of the mixture to employ (see notes) [def: [DiagonalGaussian() for i in 1:K]]

tol: Tolerance to stop the algorithm [default: 10^(-6)]

verbosity: A verbosity parameter regulating the information messages frequency [def: STD]

minimum_variance: Minimum variance for the mixtures [default: 0.05]

minimum_covariance: Minimum covariance for the mixtures with full covariance matrix [default: 0]. This should be set different than minimum_variance (see notes).

initialisation_strategy: Mixture initialisation algorithm [def: kmeans]

maximum_iterations: Maximum number of iterations [def: typemax(Int64), i.e. ∞]

rng: Random Number Generator (see FIXEDSEED) [deafult: Random.GLOBAL_RNG]

Returns:

A named touple of:

pₙₖ: Matrix of size (N x K) of the probabilities of each point i to belong to cluster j

pₖ: Probabilities of the categorical distribution (K x 1)

mixtures: Vector (K x 1) of the estimated underlying distributions

ϵ: Vector of the discrepancy (matrix norm) between pⱼₓ and the lagged pⱼₓ at each iteration

lL: The log-likelihood (without considering the last mixture optimisation)

BIC: The Bayesian Information Criterion (lower is better)

AIC: The Akaike Information Criterion (lower is better)

Notes:

The mixtures currently implemented are SphericalGaussian(μ,σ²),DiagonalGaussian(μ,σ²) and FullGaussian(μ,σ²)

Reasonable choices for the minimum_variance/Covariance depends on the mixture. For example 0.25 seems a reasonable value for the SphericalGaussian, 0.05 seems better for the DiagonalGaussian, and FullGaussian seems to prefer either very low values of variance/covariance (e.g. (0.05,0.05) ) or very big but similar ones (e.g. (100,100) ).

For initialisation_strategy, look at the documentation of init_mixtures! for the mixture you want. The provided gaussian mixtures support grid, kmeans or given. grid is faster (expecially if X contains missing values), but kmeans often provides better results.

Resources:

Paper describing gmm with missing values

Class notes from MITx 6.86x (Sec 15.9)

Limitations of gmm

Example:

julia> clusters = gmm([1 10.5;1.5 0; 1.8 8; 1.7 15; 3.2 40; 0 0; 3.3 38; 0 -2.3; 5.2 -2.4],3,verbosity=HIGH)

BetaML.GMM.init_mixtures! — Method

init_mixtures!(mixtures::Array{T,1}, X; minimum_variance=0.25, minimum_covariance=0.0, initialisation_strategy="grid",rng=Random.GLOBAL_RNG)

The parameter initialisation_strategy can be grid, kmeans or given:

grid: Uniformly cover the space observed by the data

kmeans: Use the kmeans algorithm. If the data contains missing values, a first run of predictMissing is done under init=grid to impute the missing values just to allow the kmeans algorithm. Then the em algorithm is used with the output of kmean as init values.

given: Leave the provided set of initial mixtures

BetaML.GMM.lpdf — Method

lpdf(m::DiagonalGaussian,x,mask) - Log PDF of the mixture given the observation x

BetaML.GMM.lpdf — Method

lpdf(m::FullGaussian,x,mask) - Log PDF of the mixture given the observation x

BetaML.GMM.lpdf — Method

lpdf(m::SphericalGaussian,x,mask) - Log PDF of the mixture given the observation x