The BetaML.GMM Module

BetaML.GMM — Module

GMM module

Generative (Gaussian) Mixed Model learners (supervised/unsupervised)

Provides clustering/collaborative filtering (via clustering) / missing values imputation / collaborative filtering / reccomendation systems, regressor and fitter using Generative Gaussiam Model (probabilistic).

The module provides the following functions. Use ?[function] to access their full signature and detailed documentation:

gmm(X,K;p₀,mixtures,tol,verbosity,minVariance,minCovariance,initStrategy): gmm algorithm over GMM

predictMissing(X,K;p₀,mixtures,tol,verbosity,minVariance,minCovariance): Impute mixing values ("matrix completion") using gmm as backbone. Note that this can be used for collaborative filtering / reccomendation systems often with better results than traditional algorithms as k-nearest neighbors (KNN)

{Spherical|Diagonal|Full} Gaussian mixtures are already provided. User defined mixtures can be used defining a struct as subtype of AbstractMixture and implementing for that mixture the following functions:

initMixtures!(mixtures, X; minVariance, minCovariance, initStrategy)

lpdf(m,x,mask) (for the e-step)

updateParameters!(mixtures, X, pₙₖ; minVariance, minCovariance) (the m-step)

Module Index

BetaML.GMM.GMMClusterHyperParametersSet

BetaML.GMM.GMMRegressor1

BetaML.GMM.GMMRegressor2

BetaML.GMM.estep

BetaML.GMM.gmm

BetaML.GMM.initMixtures!

BetaML.GMM.lpdf

BetaML.GMM.lpdf

BetaML.GMM.lpdf

Detailed API

BetaML.GMM.GMMClusterHyperParametersSet — Type

mutable struct GMMClusterHyperParametersSet <: BetaMLHyperParametersSet

Hyperparameters for GMM clusters and other GMM-related algorithms

Parameters:

nClasses

Number of mixtures (latent classes) to consider [def: 3]

probMixtures

Initial probabilities of the categorical distribution (nClasses x 1) [default: []]

mixtures

An array (of length K) of the mixture to employ (see notes) [def: [DiagonalGaussian() for i in 1:K]]

tol

Tolerance to stop the algorithm [default: 10^(-6)]

minVariance

Minimum variance for the mixtures [default: 0.05]

minCovariance

Minimum covariance for the mixtures with full covariance matrix [default: 0]. This should be set different than minVariance (see notes).

initStrategy

Mixture initialisation algorithm [def: kmeans]

maxIter

Maximum number of iterations [def: typemax(Int64), i.e. ∞]

BetaML.GMM.GMMRegressor1 — Type

GMMRegressor1

A multi-dimensional, missing data friendly non-linear regressor based on Generative (Gaussian) Mixture Model (strategy "1").

The training data is used to fit a probabilistic model with latent mixtures (Gaussian distributions with different covariances are already implemented) and then predictions of new data is obtained by fitting the new data to the mixtures.

For hyperparameters see GMMClusterHyperParametersSet and GMMClusterOptionsSet.

this strategy (GMMRegressor1) works by training the EM algorithm on the feature matrix X. Once the data has been probabilistically assigned to the various classes, a mean value of Y is computed for each cluster (using the probabilities as weigths). At predict time, the new data is first fitted to the learned mixtures using the e-step part of the EM algorithm to obtain the probabilistic assignment of each record to the various mixtures. Then these probabilities are multiplied to the mixture averages for the Y dimensions learned at training time to obtain the predicted value(s) for each record.

BetaML.GMM.GMMRegressor2 — Type

GMMRegressor2

A multi-dimensional, missing data friendly non-linear regressor based on Generative (Gaussian) Mixture Model.

The training data is used to fit a probabilistic model with latent mixtures (Gaussian distributions with different covariances are already implemented) and then predictions of new data is obtained by fitting the new data to the mixtures.

For hyperparameters see GMMClusterHyperParametersSet and GMMClusterOptionsSet.

Thsi strategy (GMMRegressor2) works by training the EM algorithm on a combined (hcat) matrix of X and Y. At predict time, the new data is first fitted to the learned mixtures using the e-step part of the EM algorithm (and using missing values for the dimensions belonging to Y) to obtain the probabilistic assignment of each record to the various mixtures. Thes these probabilities are multiplied to the mixture averages for the Y dimensions to obtain the predicted value(s) for each record.

BetaML.Api.fit! — Method

fit!(m::GMMClusterModel,x)

Notes:

fit! caches as record probabilities only those of the last set of data used to train the model

BetaML.Api.fit! — Method

fit!(m::GMMRegressor1,x)

Notes:

fit! caches as record probabilities only those of the last set of data used to train the model

BetaML.Api.fit! — Method

fit!(m::GMMRegressor2,x)

Notes:

fit! caches as record probabilities only those of the last set of data used to train the model

BetaML.GMM.estep — Method

estep(X,pₖ,mixtures)

E-step: assign the posterior prob p(j|xi) and computing the log-Likelihood of the parameters given the set of data(this last one for informative purposes and terminating the algorithm only)

BetaML.GMM.gmm — Method

gmm(X,K;p₀,mixtures,tol,verbosity,minVariance,minCovariance,initStrategy)

Compute Expectation-Maximisation algorithm to identify K clusters of X data, i.e. employ a Generative Mixture Model as the underlying probabilistic model.

X can contain missing values in some or all of its dimensions. In such case the learning is done only with the available data. Implemented in the log-domain for better numerical accuracy with many dimensions.

Parameters:

X : A (n x d) data to clusterise

K : Number of cluster wanted

p₀ : Initial probabilities of the categorical distribution (K x 1) [default: []]

mixtures: An array (of length K) of the mixture to employ (see notes) [def: [DiagonalGaussian() for i in 1:K]]

tol: Tolerance to stop the algorithm [default: 10^(-6)]

verbosity: A verbosity parameter regulating the information messages frequency [def: STD]

minVariance: Minimum variance for the mixtures [default: 0.05]

minCovariance: Minimum covariance for the mixtures with full covariance matrix [default: 0]. This should be set different than minVariance (see notes).

initStrategy: Mixture initialisation algorithm [def: kmeans]

maxIter: Maximum number of iterations [def: typemax(Int64), i.e. ∞]

rng: Random Number Generator (see FIXEDSEED) [deafult: Random.GLOBAL_RNG]

Returns:

A named touple of:

pₙₖ: Matrix of size (N x K) of the probabilities of each point i to belong to cluster j

pₖ: Probabilities of the categorical distribution (K x 1)

mixtures: Vector (K x 1) of the estimated underlying distributions

ϵ: Vector of the discrepancy (matrix norm) between pⱼₓ and the lagged pⱼₓ at each iteration

lL: The log-likelihood (without considering the last mixture optimisation)

BIC: The Bayesian Information Criterion (lower is better)

AIC: The Akaike Information Criterion (lower is better)

Notes:

The mixtures currently implemented are SphericalGaussian(μ,σ²),DiagonalGaussian(μ,σ²) and FullGaussian(μ,σ²)

Reasonable choices for the minVariance/Covariance depends on the mixture. For example 0.25 seems a reasonable value for the SphericalGaussian, 0.05 seems better for the DiagonalGaussian, and FullGaussian seems to prefer either very low values of variance/covariance (e.g. (0.05,0.05) ) or very big but similar ones (e.g. (100,100) ).

For initStrategy, look at the documentation of initMixtures! for the mixture you want. The provided gaussian mixtures support grid, kmeans or given. grid is faster (expecially if X contains missing values), but kmeans often provides better results.

Resources:

Paper describing gmm with missing values

Class notes from MITx 6.86x (Sec 15.9)

Limitations of gmm

Example:

julia> clusters = gmm([1 10.5;1.5 0; 1.8 8; 1.7 15; 3.2 40; 0 0; 3.3 38; 0 -2.3; 5.2 -2.4],3,verbosity=HIGH)

BetaML.GMM.initMixtures! — Method

initMixtures!(mixtures::Array{T,1}, X; minVariance=0.25, minCovariance=0.0, initStrategy="grid",rng=Random.GLOBAL_RNG)

The parameter initStrategy can be grid, kmeans or given:

grid: Uniformly cover the space observed by the data

kmeans: Use the kmeans algorithm. If the data contains missing values, a first run of predictMissing is done under init=grid to impute the missing values just to allow the kmeans algorithm. Then the em algorithm is used with the output of kmean as init values.

given: Leave the provided set of initial mixtures

BetaML.GMM.lpdf — Method

lpdf(m::DiagonalGaussian,x,mask) - Log PDF of the mixture given the observation x

BetaML.GMM.lpdf — Method

lpdf(m::FullGaussian,x,mask) - Log PDF of the mixture given the observation x

BetaML.GMM.lpdf — Method

lpdf(m::SphericalGaussian,x,mask) - Log PDF of the mixture given the observation x