The BetaML.Imputation Module

BetaML.Imputation — Module

Imputation module

Provide various imputation methods for missing data. Note that the interpretation of "missing" can be very wide. For example, reccomendation systems / collaborative filtering (e.g. suggestion of the film to watch) can well be representated as a missing data to impute problem, often with better results than traditional algorithms as k-nearest neighbors (KNN)

Provided imputers:

FeatureBasedImputer: Impute data using the feature (column) mean, optionally normalised by l-norms of the records (rows) (fastest)

GMMImputer: Impute data using a Generative (Gaussian) Mixture Model (good trade off)

RFImputer: Impute missing data using Random Forests, with optional replicable multiple imputations (most accurate).

GenericImputer: Impute missing data using a vector (one per column) of arbitrary learning models (classifiers/regressors) that implement m = Model([options]), fit!(m,X,Y) and predict(m,X).

Imputations for all these models can be optained by running mod = ImputatorModel([options]), fit!(mod,X). The data with the missing values imputed can then be obtained with predict(mod). Useinfo(m::Imputer) to retrieve further information concerning the imputation. Trained models can be also used to impute missing values in new data with predict(mox,xNew). Note that if multiple imputations are run (for the supporting imputators) predict() will return a vector of predictions rather than a single one`.

Example

julia> using Statistics, BetaML

julia> X            = [2 missing 10; 2000 4000 1000; 2000 4000 10000; 3 5 12 ; 4 8 20; 1 2 5]
6×3 Matrix{Union{Missing, Int64}}:
    2      missing     10
 2000  4000          1000
 2000  4000         10000
    3     5            12
    4     8            20
    1     2             5

julia> mod          = RFImputer(multiple_imputations=10,  rng=copy(FIXEDRNG));

julia> fit!(mod,X);

julia> vals         = predict(mod)
10-element Vector{Matrix{Union{Missing, Int64}}}:
 [2 3 10; 2000 4000 1000; … ; 4 8 20; 1 2 5]
 [2 4 10; 2000 4000 1000; … ; 4 8 20; 1 2 5]
 [2 4 10; 2000 4000 1000; … ; 4 8 20; 1 2 5]
 [2 136 10; 2000 4000 1000; … ; 4 8 20; 1 2 5]
 [2 137 10; 2000 4000 1000; … ; 4 8 20; 1 2 5]
 [2 4 10; 2000 4000 1000; … ; 4 8 20; 1 2 5]
 [2 4 10; 2000 4000 1000; … ; 4 8 20; 1 2 5]
 [2 4 10; 2000 4000 1000; … ; 4 8 20; 1 2 5]
 [2 137 10; 2000 4000 1000; … ; 4 8 20; 1 2 5]
 [2 137 10; 2000 4000 1000; … ; 4 8 20; 1 2 5]

julia> nR,nC        = size(vals[1])
(6, 3)

julia> medianValues = [median([v[r,c] for v in vals]) for r in 1:nR, c in 1:nC]
6×3 Matrix{Float64}:
    2.0     4.0     10.0
 2000.0  4000.0   1000.0
 2000.0  4000.0  10000.0
    3.0     5.0     12.0
    4.0     8.0     20.0
    1.0     2.0      5.0

julia> infos        = info(mod);

julia> infos[:n_imputed_values]
1

Module Index

BetaML.Imputation.FeatureBasedImputer

BetaML.Imputation.FeatureBasedImputerHyperParametersSet

BetaML.Imputation.GMMImputer

BetaML.Imputation.RFImputer

BetaML.Imputation.RFImputerHyperParametersSet

BetaML.Imputation.UniversalImputer

BetaML.Imputation.UniversalImputerHyperParametersSet

BetaML.Imputation.predictMissing

Detailed API

BetaML.Imputation.FeatureBasedImputer — Type

mutable struct FeatureBasedImputer <: Imputer

Simple imputer using the missing data's feature (column) statistic (def: mean), optionally normalised by l-norms of the records (rows)

Parameters:

statistics: The descriptive statistic of the column (feature) to use as imputed value [def: mean]

norm: Normalise the feature mean by l-norm norm of the records [default: nothing]. Use it (e.g. norm=1 to use the l-1 norm) if the records are highly heterogeneus (e.g. quantity exports of different countries).

Limitations:

data must be numerical

BetaML.Imputation.FeatureBasedImputerHyperParametersSet — Type

mutable struct FeatureBasedImputerHyperParametersSet <: BetaMLHyperParametersSet

Hyperparameters for the FeatureBasedImputer model

Parameters:

statistic::Function

The descriptive statistic of the column (feature) to use as imputed value [def: mean]

norm::Union{Nothing, Int64}

Normalise the feature mean by l-norm norm of the records [default: nothing]. Use it (e.g. norm=1 to use the l-1 norm) if the records are highly heterogeneus (e.g. quantity exports of different countries).

BetaML.Imputation.GMMImputer — Type

mutable struct GMMImputer <: Imputer

Missing data imputer that uses a Generative (Gaussian) Mixture Model.

For the parameters (n_classes,mixtures,..) see GMMImputerLearnableParameters.

Limitations:

data must be numerical

the resulted matrix is a Matrix{Float64}

currently the Mixtures available do not support random initialisation for missing imputation, and the rest of the algorithm (Expectation-Maximisation) is deterministic, so there is no random component involved (i.e. no multiple imputations)

BetaML.Imputation.RFImputer — Type

mutable struct RFImputer <: Imputer

Impute missing data using Random Forests, with optional replicable multiple imputations.

See RFImputerHyperParametersSet, RFHyperParametersSet and BetaMLDefaultOptionsSet for the parameters.

Notes:

Given a certain RNG and its status (e.g. RFImputer(...,rng=StableRNG(FIXEDSEED))), the algorithm is completely deterministic, i.e. replicable.

The algorithm accepts virtually any kind of data, sortable or not

BetaML.Imputation.RFImputerHyperParametersSet — Type

mutable struct RFImputerHyperParametersSet <: BetaMLHyperParametersSet

Hyperparameters for RFImputer

Parameters:

rfhpar::Any

For the underlying random forest algorithm parameters (n_trees,max_depth,min_gain,min_records,max_features:,splitting_criterion,β,initialisation_strategy, oob and rng) see RFHyperParametersSet for the specific RF algorithm parameters

forced_categorical_cols::Vector{Int64}

Specify the positions of the integer columns to treat as categorical instead of cardinal. [Default: empty vector (all numerical cols are treated as cardinal by default and the others as categorical)]

recursive_passages::Int64

Define the times to go trough the various columns to impute their data. Useful when there are data to impute on multiple columns. The order of the first passage is given by the decreasing number of missing values per column, the other passages are random [default: 1].

multiple_imputations::Int64

Determine the number of independent imputation of the whole dataset to make. Note that while independent, the imputations share the same random number generator (RNG).

Example:

julia>mod = RFImputer(n_trees=20,max_depth=10,recursive_passages=3)

BetaML.Imputation.UniversalImputer — Type

mutable struct UniversalImputer <: Imputer

Impute missing data using any regressor/classifier (not necessarily from BetaML) that implements m=Model([options]), fit!(m,X,Y) and predict(m,X)

See UniversalImputerHyperParametersSet for the hyper-parameters.

BetaML.Imputation.UniversalImputerHyperParametersSet — Type

mutable struct UniversalImputerHyperParametersSet <: BetaMLHyperParametersSet

Hyperparameters for UniversalImputer

Parameters:

estimators

Specify a regressor or classifier model (and its options/hyper-parameters) per each column of the matrix to impute. Default to random forests.

recursive_passages

Define the times to go trough the various columns to impute their data. Useful when there are data to impute on multiple columns. The order of the first passage is given by the decreasing number of missing values per column, the other passages are random [default: 1].

multiple_imputations

Determine the number of independent imputation of the whole dataset to make. Note that while independent, the imputations share the same random number generator (RNG).

BetaML.Api.fit! — Method

fit!(
    imputer::FeatureBasedImputer,
    X
) -> Union{Nothing, Matrix{Float64}}

Fit a matrix with missing data using FeatureBasedImputer

BetaML.Api.fit! — Method

fit!(m::GMMImputer, X) -> Union{Nothing, Matrix{Float64}}

Fit a matrix with missing data using GMMImputer

BetaML.Api.fit! — Method

fit!(m::RFImputer, X) -> Any

Fit a matrix with missing data using RFImputer

BetaML.Api.fit! — Method

fit!(m::UniversalImputer, X) -> Any

Fit a matrix with missing data using UniversalImputer

BetaML.Api.predict — Method

predict(m::FeatureBasedImputer, X) -> Any

Predict the missing data using the feature averages (eventually normalised) learned by fitting a FeatureBasedImputer model

BetaML.Api.predict — Method

predict(m::GMMImputer, X) -> Any

Predict the missing data using the mixtures learned by fitting a GMMImputer model

BetaML.Api.predict — Method

predict(m::RFImputer, X) -> Any

Return the data with the missing values replaced with the imputed ones using the non-linear structure learned fitting a RFImputer model.

Notes:

If multiple_imputations was set > 1 this is a vector of matrices (the individual imputations) instead of a single matrix.

BetaML.Api.predict — Method

predict(m::UniversalImputer, X) -> Any

Return the data with the missing values replaced with the imputed ones using the non-linear structure learned fitting a UniversalImputer model.

Notes:

If multiple_imputations was set > 1 this is a vector of matrices (the individual imputations) instead of a single matrix.

BetaML.Imputation.predictMissing — Function

predictMissing(X,K;initial_probmixtures,mixtures,tol,verbosity,minimum_variance,minimum_covariance)

Warning

This function is deprecated and will possibly be removed in BetaML 0.9. Use the model GMMClusterer instead.

Fill missing entries in a sparse matrix (i.e. perform a "matrix completion") assuming an underlying Gaussian Mixture probabilistic Model (GMM) fitted using an Expectation-Maximisation algorithm.

While the name of the function is predictMissing, the function can be also used for system reccomendation / collaborative filtering and GMM-based regressions. The advantage over traditional algorithms as k-nearest neighbors (KNN) is that GMM can "detect" the hidden structure of the observed data, where some observation can be similar to a certain pool of other observvations for a certain characteristic, but similar to an other pool of observations for other characteristics.

Implemented in the log-domain for better numerical accuracy with many dimensions.

Parameters:

X : A (N x D) sparse matrix of data to fill according to a GMM model

K : Number of mixtures (latent classes) to consider [def: 3]

initial_probmixtures : Initial probabilities of the categorical distribution (K x 1) [default: []]

mixtures: An array (of length K) of the mixture to employ (see notes) [def: [DiagonalGaussian() for i in 1:K]]

tol: Tolerance to stop the algorithm [default: 10^(-6)]

verbosity: A verbosity parameter regulating the information messages frequency [def: STD]

minimum_variance: Minimum variance for the mixtures [default: 0.05]

minimum_covariance: Minimum covariance for the mixtures with full covariance matrix [default: 0]. This should be set different than minimum_variance (see notes).

initialisation_strategy: Mixture initialisation algorithm [def: grid]

maximum_iterations: Maximum number of iterations [def: typemax(Int64), i.e. ∞]

rng: Random Number Generator (see FIXEDSEED) [deafult: Random.GLOBAL_RNG]

Returns:

A named touple of:

̂X̂ : The Filled Matrix of size (N x D)

nFill: The number of items filled

lL : The log-likelihood (without considering the last mixture optimisation)

BIC : The Bayesian Information Criterion (lower is better)

AIC : The Akaike Information Criterion (lower is better)

Notes:

The mixtures currently implemented are SphericalGaussian(μ,σ²),DiagonalGaussian(μ,σ²) and FullGaussian(μ,σ²)

For initialisation_strategy, look at the documentation of init_mixtures! for the mixture you want. The provided gaussian mixtures support grid, kmeans or given. grid is faster, but kmeans often provides better results.

The algorithm requires to specify a number of "latent classes" (mlixtures) to divide the dataset into. If there isn't any prior domain specific knowledge on this point one can test sevaral k and verify which one minimise the BIC or AIC criteria.

Example:

julia>  cFOut = predictMissing([1 10.5;1.5 missing; 1.8 8; 1.7 15; 3.2 40; missing missing; 3.3 38; missing -2.3; 5.2 -2.4],3)