The BetaML.Nn Module

BetaML.Nn — Module

BetaML.Nn module

Implement the functionality required to define an artificial Neural Network, train it with data, forecast data and assess its performances.

Common type of layers and optimisation algorithms are already provided, but you can define your own ones subclassing respectively the AbstractLayer and OptimisationAlgorithm abstract types.

The module provide the following types or functions. Use ?[type or function] to access their full signature and detailed documentation:

Model definition:

DenseLayer: Classical feed-forward layer with user-defined activation function

DenseNoBiasLayer: Classical layer without the bias parameter

VectorFunctionLayer: Parameterless layer whose activation function run over the ensable of its nodes rather than on each one individually

NeuralNetworkEstimator: Build the chained network and define a cost function

Each layer can use a default activation function, one of the functions provided in the Utils module (relu, tanh, softmax,...) or you can specify your own function. The derivative of the activation function can be optionally be provided, in such case training will be quicker, altought this difference tends to vanish with bigger datasets. You can alternativly implement your own layer defining a new type as subtype of the abstract type AbstractLayer. Each user-implemented layer must define the following methods:

A suitable constructor

forward(layer,x)

backward(layer,x,next_gradient)

get_params(layer)

get_gradient(layer,x,next_gradient)

set_params!(layer,w)

size(layer)

Model fitting:

fit!(nn,X,Y): fitting function

fitting_info(nn): Default callback function during fitting

SGD: The classical optimisation algorithm

ADAM: A faster moment-based optimisation algorithm

To define your own optimisation algorithm define a subtype of OptimisationAlgorithm and implement the function single_update!(θ,▽;opt_alg) and eventually init_optalg(⋅) specific for it.

Model predictions and assessment:

predict(nn) or predict(nn,X): Return the output given the data

loss(nn): Compute avg. network loss on a test set

accuracy(ŷ,y): Categorical output accuracy

While high-level functions operating on the dataset expect it to be in the standard format (nrecords × ndimensions matrices) it is customary to represent the chain of a neural network as a flow of column vectors, so all low-level operations (operating on a single datapoint) expect both the input and the output as a column vector.

Module Index

BetaML.Nn.ADAM

BetaML.Nn.DenseLayer

BetaML.Nn.DenseNoBiasLayer

BetaML.Nn.Learnable

BetaML.Nn.MultitargetNeuralNetworkRegressor

BetaML.Nn.NN

BetaML.Nn.NeuralNetworkClassifier

BetaML.Nn.NeuralNetworkEstimator

BetaML.Nn.NeuralNetworkEstimatorHyperParametersSet

BetaML.Nn.NeuralNetworkEstimatorOptionsSet

BetaML.Nn.SGD

BetaML.Nn.ScalarFunctionLayer

BetaML.Nn.VectorFunctionLayer

BetaML.Nn.backward

BetaML.Nn.buildNetwork

BetaML.Nn.forward

BetaML.Nn.get_gradient

BetaML.Nn.get_gradient

BetaML.Nn.get_gradient

BetaML.Nn.get_nparams

BetaML.Nn.get_nparams

BetaML.Nn.get_params

BetaML.Nn.get_params

BetaML.Nn.init_optalg!

BetaML.Nn.init_optalg!

BetaML.Nn.loss

BetaML.Nn.set_params!

BetaML.Nn.set_params!

BetaML.Nn.single_update!

BetaML.Nn.train!

Detailed API

BetaML.Nn.ADAM — Type

ADAM(;η, λ, β₁, β₂, ϵ)

The ADAM algorithm, an adaptive moment estimation optimiser.

Fields:

η: Learning rate (stepsize, α in the paper), as a function of the current epoch [def: t -> 0.001 (i.e. fixed)]

λ: Multiplicative constant to the learning rate [def: 1]

β₁: Exponential decay rate for the first moment estimate [range: ∈ [0,1], def: 0.9]

β₂: Exponential decay rate for the second moment estimate [range: ∈ [0,1], def: 0.999]

ϵ: Epsilon value to avoid division by zero [def: 10^-8]

BetaML.Nn.DenseLayer — Type

DenseLayer

Representation of a layer in the network

Fields:

w: Weigths matrix with respect to the input from previous layer or data (n x n pr. layer)

wb: Biases (n)

f: Activation function

df: Derivative of the activation function

BetaML.Nn.DenseNoBiasLayer — Type

DenseNoBiasLayer

Representation of a layer without bias in the network

Fields:

w: Weigths matrix with respect to the input from previous layer or data (n x n pr. layer)

f: Activation function

df: Derivative of the activation function

BetaML.Nn.Learnable — Type

Learnable(data)

Structure representing the learnable parameters of a layer or its gradient.

The learnable parameters of a layers are given in the form of a N-tuple of Array{Float64,N2} where N2 can change (e.g. we can have a layer with the first parameter being a matrix, and the second one being a scalar). We wrap the tuple on its own structure a bit for some efficiency gain, but above all to define standard mathematic operations on the gradients without doing "type pyracy" with respect to Base tuples.

BetaML.Nn.MultitargetNeuralNetworkRegressor — Type

FeedfordwardNeuralNetwork

A simple but flexible Feedforward Neural Network, from the Beta Machine Learning Toolkit (BetaML) that can be used for both regression tasks or classification ones.

Parameters:

layers

Array of layer objects [def: nothing, i.e. basic network]. See subtypes(BetaML.AbstractLayer) for supported layers

loss

Loss (cost) function [def: squared_cost].

Warning

If you change the parameter loss, you need to either provide its derivative on the parameter dloss or use autodiff with dloss=nothing.

dloss

Derivative of the loss function [def: dSquaredCost, i.e. use the derivative of the squared cost]. Use nothing for autodiff.

epochs

Number of epochs, i.e. passages trough the whole training sample [def: 1000]

batch_size

Size of each individual batch [def: 32]

opt_alg

The optimisation algorithm to update the gradient at each batch [def: ADAM()]

shuffle

Whether to randomly shuffle the data at each iteration (epoch) [def: true]

descr

An optional title and/or description for this model

cb

A call back function to provide information during training [def: fitting_info

rng

Random Number Generator (see FIXEDSEED) [deafult: Random.GLOBAL_RNG]

Notes:

data must be numerical

the label can be a n-records vector or a n-records by n-dimensions matrix (e.g. a one-hot-encoded data for classification), but the result is always a matrix.

For one-dimension regressions drop the unnecessary dimension with dropdims(ŷ,dims=2)

For classification tasks the columns should be interpreted as the probabilities for each categories

BetaML.Nn.NN — Type

NN

Representation of a Neural Network

Fields:

layers: Array of layers objects

cf: Cost function

dcf: Derivative of the cost function

trained: Control flag for trained networks

BetaML.Nn.NeuralNetworkClassifier — Type

NeuralNetworkClassifier

A simple but flexible Feedforward Neural Network, from the Beta Machine Learning Toolkit (BetaML) for classification problems.

Parameters:

layers

Array of layer objects [def: nothing, i.e. basic network]. See subtypes(BetaML.AbstractLayer) for supported layers. The last "softmax" layer is automatically added.

loss

Loss (cost) function [def: crossentropy].

Warning

If you change the parameter loss, you need to either provide its derivative on the parameter dloss or use autodiff with dloss=nothing.

dloss

Derivative of the loss function [def: dcrossentropy, i.e. the derivative of the cross-entropy]. Use nothing for autodiff.

epochs

Number of epochs, i.e. passages trough the whole training sample [def: 1000]

batch_size

Size of each individual batch [def: 32]

opt_alg

The optimisation algorithm to update the gradient at each batch [def: ADAM()]

shuffle

Whether to randomly shuffle the data at each iteration (epoch) [def: true]

descr

An optional title and/or description for this model

cb

A call back function to provide information during training [def: fitting_info

categories

The categories to represent as columns. [def: nothing, i.e. unique training values].

handle_unknown

How to handle categories not seens in training or not present in the provided categories array? "error" (default) rises an error, "infrequent" adds a specific column for these categories.

other_categories_name

Which value during prediction to assign to this "other" category (i.e. categories not seen on training or not present in the provided categories array? [def: nothing, i.e. typemax(Int64) for integer vectors and "other" for other types]. This setting is active only if handle_unknown="infrequent" and in that case it MUST be specified if Y is neither integer or strings

rng

Random Number Generator [deafult: Random.GLOBAL_RNG]

Notes:

data must be numerical

the label can be a n-records vector or a n-records by n-dimensions matrix (e.g. a one-hot-encoded data for classification), but the result is always a matrix.

For one-dimension regressions drop the unnecessary dimension with dropdims(ŷ,dims=2)

For classification tasks the columns should be interpreted as the probabilities for each categories

BetaML.Nn.NeuralNetworkEstimator — Type

NeuralNetworkEstimator

A "feedforward" neural network (supervised).

For the parameters see NeuralNetworkEstimatorLearnableParameters.

Notes:

data must be numerical

the label can be a n-records vector or a n-records by n-dimensions matrix, but the result is always a matrix.

For one-dimension regressions drop the unnecessary dimension with dropdims(ŷ,dims=2)

For classification tasks the columns should be interpreted as the probabilities for each categories

BetaML.Nn.NeuralNetworkEstimatorHyperParametersSet — Type

**`

mutable struct NeuralNetworkEstimatorHyperParametersSet <: BetaMLHyperParametersSet

`**

Hyperparameters for the Feedforward neural network model

Parameters:

layers

Array of layer objects [def: nothing, i.e. basic network]. See subtypes(BetaML.AbstractLayer) for supported layers

loss

Loss (cost) function [def: squared_cost].

Warning

If you change the parameter loss, you need to either provide its derivative on the parameter dloss or use autodiff with dloss=nothing.

dloss

Derivative of the loss function [def: dSquaredCost, i.e. use the derivative of the squared cost]. Use nothing for autodiff.

epochs

Number of epochs, i.e. passages trough the whole training sample [def: 1000]

batch_size

Size of each individual batch [def: 32]

opt_alg

The optimisation algorithm to update the gradient at each batch [def: ADAM()]

shuffle

Whether to randomly shuffle the data at each iteration (epoch) [def: true]

To know the available layers type subtypes(AbstractLayer)) and then type ?LayerName for information on how to use each layer.

BetaML.Nn.NeuralNetworkEstimatorOptionsSet — Type

NeuralNetworkEstimatorOptionsSet

A struct defining the options used by the Feedforward neural network model

Parameters:

cache

Cache the results of the fitting stage, as to allow predict(mod) [default: true]. Set it to false to save memory for large data.

descr

An optional title and/or description for this model

verbosity

The verbosity level to be used in training or prediction (see Verbosity) [deafult: STD]

cb

A call back function to provide information during training [def: fitting_info

rng

Random Number Generator (see FIXEDSEED) [deafult: Random.GLOBAL_RNG]

BetaML.Nn.SGD — Type

SGD(;η=t -> 1/(1+t), λ=2)

Stochastic Gradient Descent algorithm (default)

Fields:

η: Learning rate, as a function of the current epoch [def: t -> 1/(1+t)]

λ: Multiplicative constant to the learning rate [def: 2]

BetaML.Nn.ScalarFunctionLayer — Type

ScalarFunctionLayer

Representation of a ScalarFunction layer in the network. ScalarFunctionLayer applies the activation function directly to the output of the previous layer (i.e., without passing for a weigth matrix), but using an optional learnable parameter (an array) used as second argument, similarly to [VectorFunctionLayer(@ref). Differently from VectorFunctionLayer, the function is applied scalarwise to each node.

The number of nodes in input must be set to the same as in the previous layer

Fields:

w: Weigths (parameter) array passes as second argument to the activation function (if not empty)

n: Number of nodes in output (≡ number of nodes in input )

f: Activation function (vector)

dfx: Derivative of the (vector) activation function with respect to the layer inputs (x)

dfw: Derivative of the (vector) activation function with respect to the optional learnable weigths (w)

Notes:

The output size of this layer is the same as those of the previous layers.

BetaML.Nn.VectorFunctionLayer — Type

VectorFunctionLayer

Representation of a VectorFunction layer in the network. Vector function layer expects a vector activation function, i.e. a function taking the whole output of the previous layer an input rather than working on a single node as "normal" activation functions would do. Useful for example with the SoftMax function in classification or with the pool1D function to implement a "pool" layer in 1 dimensions. By default it is weightless, i.e. it doesn't apply any transformation to the output coming from the previous layer except the activation function. However, by passing the parameter wsize (a touple or array - tested only 1D) you can pass the learnable parameter to the activation function too. It is your responsability to be sure the activation function accept only X or also this learnable array (as second argument). The number of nodes in input must be set to the same as in the previous layer (and if you are using this for classification, to the number of classes, i.e. the previous layer must be set equal to the number of classes in the predictions).

Fields:

w: Weigths (parameter) array passes as second argument to the activation function (if not empty)

nₗ: Number of nodes in input (i.e. length of previous layer)

n: Number of nodes in output (automatically inferred in the constructor)

f: Activation function (vector)

dfx: Derivative of the (vector) activation function with respect to the layer inputs (x)

dfw: Derivative of the (vector) activation function with respect to the optional learnable weigths (w)

Notes:

The output size of this layer is given by the size of the output function,

that not necessarily is the same as the previous layers.

Base.show — Method

show(nn)

Print a representation of the Neural Network (layers, dimensions..)

Parameters:

nn: Worker network

Base.size — Method

size(layer)

Get the dimensions of the layers in terms of (dimensions in input , dimensions in output)

Notes:

You need to use import Base.size before defining this function for your layer

BetaML.Nn.backward — Method

backward(layer,x,next_gradient)

Compute backpropagation for this layer

Parameters:

layer: Worker layer

x: Input to the layer

next_gradient: Derivative of the overal loss with respect to the input of the next layer (output of this layer)

Return:

The evaluated gradient of the loss with respect to this layer inputs

BetaML.Nn.buildNetwork — Method

buildNetwork(layers,cf;dcf,name)

Instantiate a new Feedforward Neural Network

Warning

This function is deprecated and will possibly be removed in BetaML 0.9. Use the model NeuralNetworkEstimator instead.

Parameters:

layers: Array of layers objects

cf: Cost function

dcf: Derivative of the cost function [def: nothing]

name: Name of the network [def: "Neural Network"]

Notes:

Even if the network ends with a single output note, the cost function and its derivative should always expect y and ŷ as column vectors.

BetaML.Nn.forward — Method

forward(layer,x)

Predict the output of the layer given the input

Parameters:

layer: Worker layer

x: Input to the layer

Return:

An Array{T,1} of the prediction (even for a scalar)

BetaML.Nn.get_gradient — Method

get_gradient(layer,x,next_gradient)

Compute backpropagation for this layer

Parameters:

layer: Worker layer

x: Input to the layer

next_gradient: Derivative of the overaall loss with respect to the input of the next layer (output of this layer)

Return:

The evaluated gradient of the loss with respect to this layer's trainable parameters as tuple of matrices. It is up to you to decide how to organise this tuple, as long you are consistent with the get_params() and set_params() functions. Note that starting from BetaML 0.2.2 this tuple needs to be wrapped in its Learnable type.

BetaML.Nn.get_gradient — Method

get_gradient(nn,xbatch,ybatch)

Retrieve the current gradient of the weigthts (i.e. derivative of the cost with respect to the weigths)

Parameters:

nn: Worker network

xbatch: Input to the network (n,d)

ybatch: Label input (n,d)

#Notes:

The output is a vector of tuples of each layer's input weigths and bias weigths

BetaML.Nn.get_gradient — Method

get_gradient(nn,x,y)

Retrieve the current gradient of the weigthts (i.e. derivative of the cost with respect to the weigths)

Parameters:

nn: Worker network

x: Input to the network (d,1)

y: Label input (d,1)

#Notes:

The output is a vector of tuples of each layer's input weigths and bias weigths

BetaML.Nn.get_nparams — Method

get_nparams(layer)

Return the number of parameters of a layer.

It doesn't need to be implemented by each layer type, as it uses get_params().

BetaML.Nn.get_nparams — Method

get_nparams(nn) - Return the number of trainable parameters of the neural network.

BetaML.Nn.get_params — Method

get_params(layer)

Get the layers current value of its trainable parameters

Parameters:

layer: Worker layer

Return:

The current value of the layer's trainable parameters as tuple of matrices. It is up to you to decide how to organise this tuple, as long you are consistent with the get_gradient() and set_params() functions. Note that starting from BetaML 0.2.2 this tuple needs to be wrapped in its Learnable type.

BetaML.Nn.get_params — Method

get_params(nn)

Retrieve current weigthts

Parameters:

nn: Worker network

Notes:

The output is a vector of tuples of each layer's input weigths and bias weigths

BetaML.Nn.init_optalg! — Method

initoptalg!(optalg::ADAM;θ,batch_size,x,y,rng)

Initialize the ADAM algorithm with the parameters m and v as zeros and check parameter bounds

BetaML.Nn.init_optalg! — Method

initoptalg!(optalg;θ,batch_size,x,y)

Initialize the optimisation algorithm

Parameters:

opt_alg: The Optimisation algorithm to use

θ: Current parameters

batch_size: The size of the batch

x: The training (input) data

y: The training "labels" to match

rng: Random Number Generator (see FIXEDSEED) [deafult: Random.GLOBAL_RNG]

Notes:

Only a few optimizers need this function and consequently ovverride it. By default it does nothing, so if you want write your own optimizer and don't need to initialise it, you don't have to override this method

BetaML.Nn.loss — Method

loss(fnn,x,y)

Compute avg. network loss on a test set (or a single (1 × d) data point)

Parameters:

fnn: Worker network

x: Input to the network (n) or (n x d)

y: Label input (n) or (n x d)

BetaML.Nn.set_params! — Method

set_params!(layer,w)

Set the trainable parameters of the layer with the given values

Parameters:

layer: Worker layer

w: The new parameters to set (Learnable)

Notes:

The format of the tuple wrapped by Learnable must be consistent with those of the get_params() and get_gradient() functions.

BetaML.Nn.set_params! — Method

set_params!(nn,w)

Update weigths of the network

Parameters:

nn: Worker network

w: The new weights to set

BetaML.Nn.single_update! — Method

singleupdate!(θ,▽;nepoch,nbatch,batchsize,xbatch,ybatch,opt_alg)

Perform the parameters update based on the average batch gradient.

Parameters:

θ: Current parameters

▽: Average gradient of the batch

n_epoch: Count of current epoch

n_batch: Count of current batch

n_batches: Number of batches per epoch

xbatch: Data associated to the current batch

ybatch: Labels associated to the current batch

opt_alg: The Optimisation algorithm to use for the update

Notes:

This function is overridden so that each optimisation algorithm implement their

own version

Most parameters are not used by any optimisation algorithm. They are provided

to support the largest possible class of optimisation algorithms

Some optimisation algorithms may change their internal structure in this function

BetaML.Nn.train! — Method

train!(nn,x,y;epochs,batchsize,sequential,optalg,verbosity,cb)

Train a neural network with the given x,y data

Warning

This function is deprecated and will possibly be removed in BetaML 0.9. Use the model NeuralNetworkEstimator instead.

Parameters:

nn: Worker network

x: Training input to the network (records x dimensions)

y: Label input (records x dimensions)

epochs: Number of passages over the training set [def: 100]

batch_size: Size of each individual batch [def: min(size(x,1),32)]

sequential: Wether to run all data sequentially instead of random [def: false]

opt_alg: The optimisation algorithm to update the gradient at each batch [def: ADAM()]

verbosity: A verbosity parameter for the trade off information / efficiency [def: STD]

cb: A callback to provide information. [def: fitting_info]

rng: Random Number Generator (see FIXEDSEED) [deafult: Random.GLOBAL_RNG]

Return:

A named tuple with the following information

epochs: Number of epochs actually ran

ϵ_epochs: The average error on each epoch (if verbosity > LOW)

θ_epochs: The parameters at each epoch (if verbosity > STD)

Notes:

Currently supported algorithms:

SGD, the classical (Stochastic) Gradient Descent optimiser

ADAM, an adaptive moment estimation optimiser

Look at the individual optimisation algorithm (?[Name OF THE ALGORITHM]) for info on its parameter, e.g. ?SGD for the Stochastic Gradient Descent.

You can implement your own optimisation algorithm using a subtype of OptimisationAlgorithm and implementing its constructor and the update function single_update!(⋅) (type ?single_update! for details).

You can implement your own callback function, altought the one provided by default is already pretty generic (its output depends on the verbosity parameter). See fitting_info for informations on the cb parameters.

Both the callback function and the single_update! function of the optimisation algorithm can be used to stop the training algorithm, respectively returning true or stop=true.

The verbosity can be set to any of NONE,LOW,STD,HIGH,FULL.

The update is done computing the average gradient for each batch and then calling single_update! to let the optimisation algorithm perform the parameters update