The BetaML.Perceptron Module

BetaML.Perceptron — Module

Perceptron module

Provide linear and kernel classifiers.

Provide the following supervised models:

PerceptronClassifier: Train data using the classical perceptron

KernelPerceptronClassifier: Train data using the kernel perceptron

PegasosClassifier: Train data using the pegasos algorithm

All algorithms are multiclass, with PerceptronClassifier and PegasosClassifier employing a one-vs-all strategy, while KernelPerceptronClassifier employs a one-vs-one approach, and return a "probability" for each class in term of a dictionary for each record. Use mode(ŷ) to return a single class prediction per record.

These models are available in the MLJ framework as LinearPerceptron,KernelPerceptron and Pegasos respectivly.

Module Index

BetaML.Perceptron.KernelPerceptron

BetaML.Perceptron.KernelPerceptronClassifier

BetaML.Perceptron.KernelPerceptronClassifierHyperParametersSet

BetaML.Perceptron.LinearPerceptron

BetaML.Perceptron.Pegasos

BetaML.Perceptron.PegasosClassifier

BetaML.Perceptron.PegasosClassifierHyperParametersSet

BetaML.Perceptron.PerceptronClassifier

BetaML.Perceptron.PerceptronClassifierHyperParametersSet

BetaML.Perceptron.kernelPerceptron

BetaML.Perceptron.kernelPerceptronBinary

BetaML.Perceptron.pegasos

BetaML.Perceptron.pegasosBinary

BetaML.Perceptron.perceptron

BetaML.Perceptron.perceptronBinary

Detailed API

BetaML.Perceptron.KernelPerceptron — Type

The kernel perceptron algorithm using one-vs-one for multiclass, from the Beta Machine Learning Toolkit (BetaML).

BetaML.Perceptron.KernelPerceptronClassifier — Type

mutable struct KernelPerceptronClassifier <: BetaMLSupervisedModel

A "kernel" version of the Perceptron model (supervised) with user configurable kernel function.

For the parameters see ?KernelPerceptronClassifierHyperParametersSet and ?BetaMLDefaultOptionsSet

Limitations:

data must be numerical

online training (retraining) is not supported

BetaML.Perceptron.KernelPerceptronClassifierHyperParametersSet — Type

mutable struct KernelPerceptronClassifierHyperParametersSet <: BetaMLHyperParametersSet

Hyperparameters for the KernelPerceptronClassifier model

Parameters:

kernel

Kernel function to employ. See ?radial_kernel or ?polynomial_kernel for details or check ?BetaML.Utils to verify if other kernels are defined (you can alsways define your own kernel) [def: radial_kernel]

initial_errors

Initial distribution of the number of errors errors [def: nothing, i.e. zeros]. If provided, this should be a nModels-lenght vector of nRecords integer values vectors , where nModels is computed as (n_classes * (n_classes - 1)) / 2

epochs

Maximum number of epochs, i.e. passages trough the whole training sample [def: 100]

shuffle

Whether to randomly shuffle the data at each iteration (epoch) [def: false]

tunemethod

The method - and its parameters - to employ for hyperparameters autotuning. See SuccessiveHalvingSearch for the default method. To implement automatic hyperparameter tuning during the (first) fit! call simply set autotune=true and eventually change the default tunemethod options (including the parameter ranges, the resources to employ and the loss function to adopt).

BetaML.Perceptron.LinearPerceptron — Type

The classical perceptron algorithm using one-vs-all for multiclass, from the Beta Machine Learning Toolkit (BetaML).

BetaML.Perceptron.Pegasos — Type

The gradient-based linear "pegasos" classifier using one-vs-all for multiclass, from the Beta Machine Learning Toolkit (BetaML).

BetaML.Perceptron.PegasosClassifier — Type

mutable struct PegasosClassifier <: BetaMLSupervisedModel

The PegasosClassifier model, a linear, gradient-based classifier. Multiclass is supported using a one-vs-all approach.

See ?PegasosClassifierHyperParametersSet and ?BetaMLDefaultOptionsSet for applicable hyperparameters and options.

BetaML.Perceptron.PegasosClassifierHyperParametersSet — Type

mutable struct PegasosClassifierHyperParametersSet <: BetaMLHyperParametersSet

Hyperparameters for the PegasosClassifier model.

Parameters:

learning_rate::Function

Learning rate [def: (epoch -> 1/sqrt(epoch))]

learning_rate_multiplicative::Float64

Multiplicative term of the learning rate [def: 0.5]

initial_parameters::Union{Nothing, Matrix{Float64}}

Initial parameters. If given, should be a matrix of n-classes by feature dimension + 1 (to include the constant term as the first element) [def: nothing, i.e. zeros]

epochs::Int64

Maximum number of epochs, i.e. passages trough the whole training sample [def: 1000]

shuffle::Bool

Whether to randomly shuffle the data at each iteration (epoch) [def: false]

force_origin::Bool

Whether to force the parameter associated with the constant term to remain zero [def: false]

return_mean_hyperplane::Bool

Whether to return the average hyperplane coefficients instead of the final ones [def: false]

tunemethod::AutoTuneMethod

The method - and its parameters - to employ for hyperparameters autotuning. See SuccessiveHalvingSearch for the default method. To implement automatic hyperparameter tuning during the (first) fit! call simply set autotune=true and eventually change the default tunemethod options (including the parameter ranges, the resources to employ and the loss function to adopt).

BetaML.Perceptron.PerceptronClassifier — Type

mutable struct PerceptronClassifier <: BetaMLSupervisedModel

The classical "perceptron" linear classifier (supervised).

For the parameters see ?PerceptronClassifierHyperParametersSet and ?BetaMLDefaultOptionsSet.

Notes:

data must be numerical

online fitting (re-fitting with new data) is not supported

BetaML.Perceptron.PerceptronClassifierHyperParametersSet — Type

mutable struct PerceptronClassifierHyperParametersSet <: BetaMLHyperParametersSet

Hyperparameters for the PerceptronClassifier model

Parameters:

initial_parameters::Union{Nothing, Matrix{Float64}}

Initial parameters. If given, should be a matrix of n-classes by feature dimension + 1 (to include the constant term as the first element) [def: nothing, i.e. zeros]

epochs::Int64

Maximum number of epochs, i.e. passages trough the whole training sample [def: 1000]

shuffle::Bool

Whether to randomly shuffle the data at each iteration (epoch) [def: false]

force_origin::Bool

Whether to force the parameter associated with the constant term to remain zero [def: false]

return_mean_hyperplane::Bool

Whether to return the average hyperplane coefficients instead of the final ones [def: false]

tunemethod::AutoTuneMethod

The method - and its parameters - to employ for hyperparameters autotuning. See SuccessiveHalvingSearch for the default method. To implement automatic hyperparameter tuning during the (first) fit! call simply set autotune=true and eventually change the default tunemethod options (including the parameter ranges, the resources to employ and the loss function to adopt).

BetaML.Api.predict — Function

predict(x,θ,θ₀)

Predict a binary label {-1,1} given the feature vector and the linear coefficients

Warning

This function is deprecated and will possibly be removed in BetaML 0.9. Use the predict function with your desired model instead.

Parameters:

x: Feature matrix of the training data (n × d)

θ: The trained parameters

θ₀: The trained bias barameter [def: 0]

Return :

y: Vector of the predicted labels

Example:

julia> predict([1.1 2.1; 5.3 4.2; 1.8 1.7], [3.2,1.2])

BetaML.Api.predict — Method

predict(x,xtrain,ytrain,α;K)

Predict a binary label {-1,1} given the feature vector and the training data together with their errors (as trained by a kernel perceptron algorithm)

Warning

This function is deprecated and will possibly be removed in BetaML 0.9. Use the predict function with your desired model instead.

Parameters:

x: Feature matrix of the data to predict (n × d)

xtrain: The feature vectors used for the training

ytrain: The labels of the training set

α: The errors associated to each record

K: The kernel function used for the training and to be used for the prediction [def: radial_kernel]

Return :

y: Vector of the predicted labels

Example:

julia> predict([1.1 2.1; 5.3 4.2; 1.8 1.7], [3.2,1.2])

BetaML.Api.predict — Method

predict(m::KernelPerceptronClassifier, X) -> Any

Predict labels using a fitted KernelPerceptronClassifier model.

BetaML.Api.predict — Method

predict(m::PegasosClassifier, X) -> Any

Predict labels using a fitted PegasosClassifier model.

BetaML.Api.predict — Method

predict(m::PerceptronClassifier, X) -> Any

Predict the labels associated to some feature data using the linear coefficients learned by fitting a PerceptronClassifier model

BetaML.Api.predict — Method

predict(x,xtrain,ytrain,α,classes;K)

Predict a multiclass label given the new feature vector and a trained kernel perceptron model.

Warning

This function is deprecated and will possibly be removed in BetaML 0.9. Use the predict function with your desired model instead.

Parameters:

x: Feature matrix of the data to predict (n × d)

xtrain: A vector of the feature matrix used for training each of the one-vs-one class matches (i.e. model.x)

ytrain: A vector of the label vector used for training each of the one-vs-one class matches (i.e. model.y)

α: A vector of the errors associated to each record (i.e. model.α)

classes: The overal classes encountered in training (i.e. model.classes)

K: The kernel function used for the training and to be used for the prediction [def: radial_kernel]

Return :

ŷ: Vector of dictionaries label=>probability (warning: it isn't really a probability, it is just the standardized number of matches "won" by this class compared with the other classes)

Notes:

Use mode(ŷ) if you want a single predicted label per record

Example:

julia> model  = kernelPerceptron([1.1 2.1; 5.3 4.2; 1.8 1.7], [-1,1,-1])
julia> ŷtrain = Perceptron.predict([10 10; 2.2 2.5],model.x,model.y,model.α, model.classes,K=model.K)

BetaML.Api.predict — Method

predict(x,θ,θ₀,classes)

Predict a multiclass label given the feature vector, the linear coefficients and the classes vector

Warning

This function is deprecated and will possibly be removed in BetaML 0.9. Use the predict function of your desired model instead.

Parameters:

x: Feature matrix of the training data (n × d)

θ: Vector of the trained parameters for each one-vs-all model (i.e. model.θ)

θ₀: Vector of the trained bias barameter for each one-vs-all model (i.e. model.θ₀)

classes: The overal classes encountered in training (i.e. model.classes)

Return :

ŷ: Vector of dictionaries label=>probability

Notes:

Use mode(ŷ) if you want a single predicted label per record

Example:

```julia julia> model = perceptron([1.1 2.1; 5.3 4.2; 1.8 1.7], [-1,1,-1]) julia> ŷtrain = predict([10 10; 2.5 2.5],model.θ,model.θ₀, model.classes)

BetaML.Perceptron.kernelPerceptron — Method

kernelPerceptron(x,y;K,T,α,nMsgs,shuffle)

Train a multiclass kernel classifier "perceptron" algorithm based on x and y.

Warning

This function is deprecated and will possibly be removed in BetaML 0.9. Use the model kernelPerceptron() instead.

kernelPerceptron is a (potentially) non-linear perceptron-style classifier employing user-defined kernel funcions. Multiclass is supported using a one-vs-one approach.

Parameters:

x: Feature matrix of the training data (n × d)

y: Associated labels of the training data

K: Kernel function to employ. See ?radial_kernel or ?polynomial_kernelfor details or check ?BetaML.Utils to verify if other kernels are defined (you can alsways define your own kernel) [def: radial_kernel]

T: Maximum number of iterations (aka "epochs") across the whole set (if the set is not fully classified earlier) [def: 100]

α: Initial distribution of the number of errors errors [def: nothing, i.e. zeros]. If provided, this should be a nModels-lenght vector of nRecords integer values vectors , where nModels is computed as (n_classes * (n_classes - 1)) / 2

nMsg: Maximum number of messages to show if all iterations are done [def: 0]

shuffle: Whether to randomly shuffle the data at each iteration [def: false]

rng: Random Number Generator (see FIXEDSEED) [deafult: Random.GLOBAL_RNG]

Return a named tuple with:

x: The x data (eventually shuffled if shuffle=true)

y: The label

α: The errors associated to each record

classes: The labels classes encountered in the training

Notes:

The trained model can then be used to make predictions using the function predict().

This model is available in the MLJ framework as the KernelPerceptron

Example:

julia> model = kernelPerceptron([1.1 1.1; 5.3 4.2; 1.8 1.7; 7.5 5.2;], ["a","c","b","c"])
julia> ŷtest = Perceptron.predict([10 10; 2.2 2.5; 1 1],model.x,model.y,model.α, model.classes,K=model.K)

BetaML.Perceptron.kernelPerceptronBinary — Method

kernelPerceptronBinary(x,y;K,T,α,nMsgs,shuffle)

Train a binary kernel classifier "perceptron" algorithm based on x and y

Warning

This function is deprecated and will possibly be removed in BetaML 0.9. Use the model KernelPerceptronClassifier() instead.

Parameters:

x: Feature matrix of the training data (n × d)

y: Associated labels of the training data, in the format of ⨦ 1

K: Kernel function to employ. See ?radial_kernel or ?polynomial_kernelfor details or check ?BetaML.Utils to verify if other kernels are defined (you can alsways define your own kernel) [def: radial_kernel]

T: Maximum number of iterations across the whole set (if the set is not fully classified earlier) [def: 1000]

α: Initial distribution of the errors [def: zeros(length(y))]

nMsg: Maximum number of messages to show if all iterations are done

shuffle: Whether to randomly shuffle the data at each iteration [def: false]

rng: Random Number Generator (see FIXEDSEED) [deafult: Random.GLOBAL_RNG]

Return a named tuple with:

x: the x data (eventually shuffled if shuffle=true)

y: the label

α: the errors associated to each record

errors: the number of errors in the last iteration

besterrors: the minimum number of errors in classifying the data ever reached

iterations: the actual number of iterations performed

separated: a flag if the data has been successfully separated

Notes:

The trained data can then be used to make predictions using the function predict(). If the option shuffle has been used, it is important to use there the returned (x,y,α) as these would have been shuffled compared with the original (x,y).

Please see @kernelPerceptron for a multi-class version

Example:

julia> model = kernelPerceptronBinary([1.1 2.1; 5.3 4.2; 1.8 1.7], [-1,1,-1])

BetaML.Perceptron.pegasos — Method

pegasos(x,y;θ,θ₀,λ,η,T,nMsgs,shuffle,forceorigin,returnmean_hyperplane)

Train the multiclass classifier "pegasos" algorithm according to x (features) and y (labels)

Warning

This function is deprecated and will possibly be removed in BetaML 0.9. Use the model PegasosClassifier instead.

PegasosClassifier is a linear, gradient-based classifier. Multiclass is supported using a one-vs-all approach.

Parameters:

x: Feature matrix of the training data (n × d)

y: Associated labels of the training data, can be in any format (string, integers..)

θ: Initial value of the weights (parameter) [def: zeros(d)]

θ₀: Initial value of the weight (parameter) associated to the constant term [def: 0]

λ: Multiplicative term of the learning rate

η: Learning rate [def: (t -> 1/sqrt(t))]

T: Maximum number of iterations across the whole set (if the set is not fully classified earlier) [def: 1000]

nMsg: Maximum number of messages to show if all iterations are done

shuffle: Whether to randomly shuffle the data at each iteration [def: false]

force_origin: Whether to force θ₀ to remain zero [def: false]

return_mean_hyperplane: Whether to return the average hyperplane coefficients instead of the average ones [def: false]

rng: Random Number Generator (see FIXEDSEED) [deafult: Random.GLOBAL_RNG]

Return a named tuple with:

θ: The weights of the classifier

θ₀: The weight of the classifier associated to the constant term

classes: The classes (unique values) of y

Notes:

The trained parameters can then be used to make predictions using the function predict().

This model is available in the MLJ framework as the Pegasos

Example:

julia> model = pegasos([1.1 2.1; 5.3 4.2; 1.8 1.7], [-1,1,-1])
julia> ŷ     = predict([2.1 3.1; 7.3 5.2], model.θ, model.θ₀, model.classes)

BetaML.Perceptron.pegasosBinary — Method

pegasosBinary(x,y;θ,θ₀,λ,η,T,nMsgs,shuffle,force_origin)

Train the peagasos algorithm based on x and y (labels)

Warning

This function is deprecated and will possibly be removed in BetaML 0.9. Use the model PegasosClassifier instead.

Parameters:

x: Feature matrix of the training data (n × d)

y: Associated labels of the training data, in the format of ⨦ 1

θ: Initial value of the weights (parameter) [def: zeros(d)]

θ₀: Initial value of the weight (parameter) associated to the constant term [def: 0]

λ: Multiplicative term of the learning rate

η: Learning rate [def: (t -> 1/sqrt(t))]

T: Maximum number of iterations across the whole set (if the set is not fully classified earlier) [def: 1000]

nMsg: Maximum number of messages to show if all iterations are done

shuffle: Whether to randomly shuffle the data at each iteration [def: false]

force_origin: Whether to force θ₀ to remain zero [def: false]

Return a named tuple with:

θ: The final weights of the classifier

θ₀: The final weight of the classifier associated to the constant term

avgθ: The average weights of the classifier

avgθ₀: The average weight of the classifier associated to the constant term

errors: The number of errors in the last iteration

besterrors: The minimum number of errors in classifying the data ever reached

iterations: The actual number of iterations performed

separated: Weather the data has been successfully separated

Notes:

The trained parameters can then be used to make predictions using the function predict().

Example:

julia> pegasos([1.1 2.1; 5.3 4.2; 1.8 1.7], [-1,1,-1])

BetaML.Perceptron.perceptron — Method

perceptron(x,y;θ,θ₀,T,nMsgs,shuffle,force_origin,return_mean_hyperplane)

Train the multiclass classifier "perceptron" algorithm based on x and y (labels).

Warning

This function is deprecated and will possibly be removed in BetaML 0.9. Use the model PerceptronClassifier instead.

The perceptron is a linear classifier. Multiclass is supported using a one-vs-all approach.

Parameters:

x: Feature matrix of the training data (n × d)

y: Associated labels of the training data, can be in any format (string, integers..)

θ: Initial value of the weights (parameter) [def: zeros(d)]

θ₀: Initial value of the weight (parameter) associated to the constant term [def: 0]

T: Maximum number of iterations across the whole set (if the set is not fully classified earlier) [def: 1000]

nMsg: Maximum number of messages to show if all iterations are done [def: 0]

shuffle: Whether to randomly shuffle the data at each iteration [def: false]

force_origin: Whether to force θ₀ to remain zero [def: false]

return_mean_hyperplane: Whether to return the average hyperplane coefficients instead of the final ones [def: false]

rng: Random Number Generator (see FIXEDSEED) [deafult: Random.GLOBAL_RNG]

Return a named tuple with:

θ: The weights of the classifier

θ₀: The weight of the classifier associated to the constant term

classes: The classes (unique values) of y

Notes:

The trained parameters can then be used to make predictions using the function predict().

This model is available in the MLJ framework as the LinearPerceptron

Example:

julia> model = perceptron([1.1 2.1; 5.3 4.2; 1.8 1.7], [-1,1,-1])
julia> ŷ     = predict([2.1 3.1; 7.3 5.2], model.θ, model.θ₀, model.classes)

BetaML.Perceptron.perceptronBinary — Method

perceptronBinary(x,y;θ,θ₀,T,nMsgs,shuffle,force_origin)

Warning

This function is deprecated and will possibly be removed in BetaML 0.9. Use the model PerceptronClassifier() instead.

Train the binary classifier "perceptron" algorithm based on x and y (labels)

Parameters:

x: Feature matrix of the training data (n × d)

y: Associated labels of the training data, in the format of ⨦ 1

θ: Initial value of the weights (parameter) [def: zeros(d)]

θ₀: Initial value of the weight (parameter) associated to the constant term [def: 0]

T: Maximum number of iterations across the whole set (if the set is not fully classified earlier) [def: 1000]

nMsg: Maximum number of messages to show if all iterations are done

shuffle: Whether to randomly shuffle the data at each iteration [def: false]

force_origin: Whether to force θ₀ to remain zero [def: false]

rng: Random Number Generator (see FIXEDSEED) [deafult: Random.GLOBAL_RNG]

Return a named tuple with:

θ: The final weights of the classifier

θ₀: The final weight of the classifier associated to the constant term

avgθ: The average weights of the classifier

avgθ₀: The average weight of the classifier associated to the constant term

errors: The number of errors in the last iteration

besterrors: The minimum number of errors in classifying the data ever reached

iterations: The actual number of iterations performed

separated: Weather the data has been successfully separated

Notes:

The trained parameters can then be used to make predictions using the function predict().

Example:

julia> model = perceptronBinary([1.1 2.1; 5.3 4.2; 1.8 1.7], [-1,1,-1])