The BetaML.Trees Module

BetaML.Trees — Module

BetaML.Trees module

Implement the DecisionTreeEstimator and RandomForestEstimator models (Decision Trees and Random Forests).

Both Decision Trees and Random Forests can be used for regression or classification problems, based on the type of the labels (numerical or not). The automatic selection can be overridden with the parameter force_classification=true, typically if labels are integer representing some categories rather than numbers. For classification problems the output of predict is a dictionary with the key being the labels with non-zero probabilitity and the corresponding value its probability; for regression it is a numerical value.

Please be aware that, differently from most other implementations, the Random Forest algorithm collects and averages the probabilities from the trees, rather than just repording the mode, i.e. no information is lost and the output of the forest classifier is still a PMF.

To retrieve the prediction with the highest probability use mode over the prediciton returned by the model. Most error/accuracy measures in the Util BetaML module works diretly with this format.

Missing data and trully unordered types are supported on the features, both on training and on prediction.

The module provide the following functions. Use ?[type or function] to access their full signature and detailed documentation:

Features are expected to be in the standard format (nRecords × nDimensions matrices) and the labels (either categorical or numerical) as a nRecords column vector.

Acknowlegdments: originally based on the Josh Gordon's code

Module Index

BetaML.Trees.DTHyperParametersSet

BetaML.Trees.DecisionTreeClassifier

BetaML.Trees.DecisionTreeEstimator

BetaML.Trees.DecisionTreeRegressor

BetaML.Trees.InfoNode

BetaML.Trees.RFHyperParametersSet

BetaML.Trees.RandomForestClassifier

BetaML.Trees.RandomForestEstimator

BetaML.Trees.RandomForestRegressor

BetaML.Trees.buildForest

BetaML.Trees.buildTree

BetaML.Trees.wrap

Detailed API

BetaML.Trees.DTHyperParametersSet — Type

mutable struct DTHyperParametersSet <: BetaMLHyperParametersSet

Hyperparameters for DecisionTreeEstimator (Decision Tree).

Parameters:

max_depth::Union{Nothing, Int64}

The maximum depth the tree is allowed to reach. When this is reached the node is forced to become a leaf [def: nothing, i.e. no limits]

min_gain::Float64

The minimum information gain to allow for a node's partition [def: 0]

min_records::Int64

The minimum number of records a node must holds to consider for a partition of it [def: 2]

max_features::Union{Nothing, Int64}

The maximum number of (random) features to consider at each partitioning [def: nothing, i.e. look at all features]

force_classification::Bool

Whether to force a classification task even if the labels are numerical (typically when labels are integers encoding some feature rather than representing a real cardinal measure) [def: false]

splitting_criterion::Union{Nothing, Function}

This is the name of the function to be used to compute the information gain of a specific partition. This is done by measuring the difference betwwen the "impurity" of the labels of the parent node with those of the two child nodes, weighted by the respective number of items. [def: nothing, i.e. gini for categorical labels (classification task) and variance for numerical labels(regression task)]. Either gini, entropy, variance or a custom function. It can also be an anonymous function.

BetaML.Trees.DecisionTreeClassifier — Type

mutable struct DecisionTreeClassifier <: MLJModelInterface.Probabilistic

A simple Decision Tree for classification with support for Missing data, from the Beta Machine Learning Toolkit (BetaML).

Hyperparameters:

max_depth::Int64

The maximum depth the tree is allowed to reach. When this is reached the node is forced to become a leaf [def: 0, i.e. no limits]

min_gain::Float64

The minimum information gain to allow for a node's partition [def: 0]

min_records::Int64

The minimum number of records a node must holds to consider for a partition of it [def: 2]

max_features::Int64

The maximum number of (random) features to consider at each partitioning [def: 0, i.e. look at all features]

splitting_criterion::Function

This is the name of the function to be used to compute the information gain of a specific partition. This is done by measuring the difference betwwen the "impurity" of the labels of the parent node with those of the two child nodes, weighted by the respective number of items. [def: gini]. Either gini, entropy or a custom function. It can also be an anonymous function.

rng::Random.AbstractRNG

A Random Number Generator to be used in stochastic parts of the code [deafult: Random.GLOBAL_RNG]

BetaML.Trees.DecisionTreeEstimator — Type

mutable struct DecisionTreeEstimator <: BetaMLSupervisedModel

A Decision Tree classifier and regressor (supervised).

Decision Tree works by finding the "best" question to split the fitting data (according to the metric specified by the parameter splitting_criterion on the associated labels) untill either all the dataset is separated or a terminal condition is reached.

For the parameters see ?DTHyperParametersSet and ?BetaMLDefaultOptionsSet.

Notes:

Online fitting (re-fitting with new data) is not supported

Missing data (in the feature dataset) is supported.

BetaML.Trees.DecisionTreeRegressor — Type

mutable struct DecisionTreeRegressor <: MLJModelInterface.Deterministic

A simple Decision Tree for regression with support for Missing data, from the Beta Machine Learning Toolkit (BetaML).

Hyperparameters:

max_depth::Int64

The maximum depth the tree is allowed to reach. When this is reached the node is forced to become a leaf [def: 0, i.e. no limits]

min_gain::Float64

The minimum information gain to allow for a node's partition [def: 0]

min_records::Int64

The minimum number of records a node must holds to consider for a partition of it [def: 2]

max_features::Int64

The maximum number of (random) features to consider at each partitioning [def: 0, i.e. look at all features]

splitting_criterion::Function

This is the name of the function to be used to compute the information gain of a specific partition. This is done by measuring the difference betwwen the "impurity" of the labels of the parent node with those of the two child nodes, weighted by the respective number of items. [def: variance]. Either variance or a custom function. It can also be an anonymous function.

rng::Random.AbstractRNG

A Random Number Generator to be used in stochastic parts of the code [deafult: Random.GLOBAL_RNG]

BetaML.Trees.InfoNode — Type

These types are introduced so that additional information currently not present in a DecisionTree-structure – namely the feature names – can be used for visualization.

BetaML.Trees.RFHyperParametersSet — Type

mutable struct RFHyperParametersSet <: BetaMLHyperParametersSet

Hyperparameters for RandomForestEstimator (Random Forest).

Parameters:

n_trees::Int64

Number of (decision) trees in the forest [def: 30]

max_depth::Union{Nothing, Int64}

The maximum depth the tree is allowed to reach. When this is reached the node is forced to become a leaf [def: nothing, i.e. no limits]

min_gain::Float64

The minimum information gain to allow for a node's partition [def: 0]

min_records::Int64

The minimum number of records a node must holds to consider for a partition of it [def: 2]

max_features::Union{Nothing, Int64}

The maximum number of (random) features to consider when choosing the optimal partition of the dataset [def: nothing, i.e. square root of the dimensions of the training data`]

force_classification::Bool

Whether to force a classification task even if the labels are numerical (typically when labels are integers encoding some feature rather than representing a real cardinal measure) [def: false]

splitting_criterion::Union{Nothing, Function}

Either gini, entropy or variance. This is the name of the function to be used to compute the information gain of a specific partition. This is done by measuring the difference betwwen the "impurity" of the labels of the parent node with those of the two child nodes, weighted by the respective number of items. [def: nothing, i.e. gini for categorical labels (classification task) and variance for numerical labels(regression task)]. It can be an anonymous function.

beta::Float64

Parameter that regulate the weights of the scoring of each tree, to be (optionally) used in prediction based on the error of the individual trees computed on the records on which trees have not been trained. Higher values favour "better" trees, but too high values will cause overfitting [def: 0, i.e. uniform weigths]

oob::Bool

Wheter to compute the Out-Of-Bag error, an estimation of the validation error (the mismatching error for classification and the relative mean error for regression jobs).

BetaML.Trees.RandomForestClassifier — Type

mutable struct RandomForestClassifier <: MLJModelInterface.Probabilistic

A simple Random Forest for classification with support for Missing data, from the Beta Machine Learning Toolkit (BetaML).

Hyperparameters:

n_trees::Int64

max_depth::Int64

The maximum depth the tree is allowed to reach. When this is reached the node is forced to become a leaf [def: 0, i.e. no limits]

min_gain::Float64

The minimum information gain to allow for a node's partition [def: 0]

min_records::Int64

The minimum number of records a node must holds to consider for a partition of it [def: 2]

max_features::Int64

The maximum number of (random) features to consider at each partitioning [def: 0, i.e. square root of the data dimensions]

splitting_criterion::Function

This is the name of the function to be used to compute the information gain of a specific partition. This is done by measuring the difference betwwen the "impurity" of the labels of the parent node with those of the two child nodes, weighted by the respective number of items. [def: gini]. Either gini, entropy or a custom function. It can also be an anonymous function.

β::Float64

Parameter that regulate the weights of the scoring of each tree, to be (optionally) used in prediction based on the error of the individual trees computed on the records on which trees have not been trained. Higher values favour "better" trees, but too high values will cause overfitting [def: 0, i.e. uniform weigths]

rng::Random.AbstractRNG

A Random Number Generator to be used in stochastic parts of the code [deafult: Random.GLOBAL_RNG]

BetaML.Trees.RandomForestEstimator — Type

mutable struct RandomForestEstimator <: BetaMLSupervisedModel

A Random Forest classifier and regressor (supervised).

Random forests are ensemble of Decision Trees models (see ?DecisionTreeEstimator).

For the parameters see ?RFHyperParametersSet and ?BetaMLDefaultOptionsSet.

Notes :

Each individual decision tree is built using bootstrap over the data, i.e. "sampling N records with replacement" (hence, some records appear multiple times and some records do not appear in the specific tree training). The maxx_feature injects further variability and reduces the correlation between the forest trees.

The predictions of the "forest" (using the function predict()) are then the aggregated predictions of the individual trees (from which the name "bagging": boostrap aggregating).

The performances of each individual trees, as measured using the records they have not being trained with, can then be (optionally) used as weights in the predict function. The parameter beta ≥ 0 regulate the distribution of these weights: larger is β, the greater the importance (hence the weights) attached to the best-performing trees compared to the low-performing ones. Using these weights can significantly improve the forest performances (especially using small forests), however the correct value of beta depends on the problem under exam (and the chosen caratteristics of the random forest estimator) and should be cross-validated to avoid over-fitting.

Note that training RandomForestEstimator uses multiple threads if these are available. You can check the number of threads available with Threads.nthreads(). To set the number of threads in Julia either set the environmental variable JULIA_NUM_THREADS (before starting Julia) or start Julia with the command line option --threads (most integrated development editors for Julia already set the number of threads to 4).

Online fitting (re-fitting with new data) is not supported

Missing data (in the feature dataset) is supported.

BetaML.Trees.RandomForestRegressor — Type

mutable struct RandomForestRegressor <: MLJModelInterface.Deterministic

A simple Random Forest for regression with support for Missing data, from the Beta Machine Learning Toolkit (BetaML).

Hyperparameters:

n_trees::Int64

max_depth::Int64

The maximum depth the tree is allowed to reach. When this is reached the node is forced to become a leaf [def: 0, i.e. no limits]

min_gain::Float64

The minimum information gain to allow for a node's partition [def: 0]

min_records::Int64

The minimum number of records a node must holds to consider for a partition of it [def: 2]

max_features::Int64

The maximum number of (random) features to consider at each partitioning [def: 0, i.e. square root of the data dimension]

splitting_criterion::Function

This is the name of the function to be used to compute the information gain of a specific partition. This is done by measuring the difference betwwen the "impurity" of the labels of the parent node with those of the two child nodes, weighted by the respective number of items. [def: variance]. Either variance or a custom function. It can also be an anonymous function.

β::Float64

Parameter that regulate the weights of the scoring of each tree, to be (optionally) used in prediction based on the error of the individual trees computed on the records on which trees have not been trained. Higher values favour "better" trees, but too high values will cause overfitting [def: 0, i.e. uniform weigths]

rng::Random.AbstractRNG

A Random Number Generator to be used in stochastic parts of the code [deafult: Random.GLOBAL_RNG]

BetaML.Trees.buildForest — Method

buildForest(x, y, ntrees; maxdepth, mingain, minrecords, maxfeatures, splittingcriterion, force_classification)

Builds (define and train) a "forest" of Decision Trees.

Warning

This function is deprecated and will possibly be removed in BetaML 0.9. Use RandomForestEstimator instead.

Parameters:

See buildTree. The function has all the parameters of bildTree (with the max_features defaulting to √D instead of D) plus the following parameters:

n_trees: Number of trees in the forest [def: 30]

β: Parameter that regulate the weights of the scoring of each tree, to be (optionally) used in prediction (see later) [def: 0, i.e. uniform weigths]

oob: Whether to coompute the out-of-bag error, an estimation of the generalization accuracy [def: false]

rng: Random Number Generator (see FIXEDSEED) [deafult: Random.GLOBAL_RNG]

Output:

The function returns a Forest object (see Forest).

The forest weights default to array of ones if β ≤ 0 and the oob error to +Inf if oob == false.

Notes :

Each individual decision tree is built using bootstrap over the data, i.e. "sampling N records with replacement" (hence, some records appear multiple times and some records do not appear in the specific tree training). The maxFeature injects further variability and reduces the correlation between the forest trees.

The predictions of the "forest" (using the function predict()) are then the aggregated predictions of the individual trees (from which the name "bagging": boostrap aggregating).

This function optionally reports a weight distribution of the performances of eanch individual trees, as measured using the records he has not being trained with. These weights can then be (optionally) used in the predict function. The parameter β ≥ 0 regulate the distribution of these weights: larger is β, the greater the importance (hence the weights) attached to the best-performing trees compared to the low-performing ones. Using these weights can significantly improve the forest performances (especially using small forests), however the correct value of β depends on the problem under exam (and the chosen caratteristics of the random forest estimator) and should be cross-validated to avoid over-fitting.

Note that this function uses multiple threads if these are available. You can check the number of threads available with Threads.nthreads(). To set the number of threads in Julia either set the environmental variable JULIA_NUM_THREADS (before starting Julia) or start Julia with the command line option --threads (most integrated development editors for Julia already set the number of threads to 4).

BetaML.Trees.buildTree — Method

buildTree(x, y, depth; maxdepth, mingain, minrecords, maxfeatures, splittingcriterion, forceclassification)

Builds (define and train) a Decision Tree.

Warning

This function is deprecated and will possibly be removed in BetaML 0.9. Use DecisionTreeEstimator instead.

Given a dataset of features x and the corresponding dataset of labels y, recursivelly build a decision tree by finding at each node the best question to split the data untill either all the dataset is separated or a terminal condition is reached. The given tree is then returned.

Parameters:

x: The dataset's features (N × D)

y: The dataset's labels (N × 1)

max_depth: The maximum depth the tree is allowed to reach. When this is reached the node is forced to become a leaf [def: N, i.e. no limits]

min_gain: The minimum information gain to allow for a node's partition [def: 0]

min_records: The minimum number of records a node must holds to consider for a partition of it [def: 2]

max_features: The maximum number of (random) features to consider at each partitioning [def: D, i.e. look at all features]

splitting_criterion: Either gini, entropy or variance (see infoGain ) [def: gini for categorical labels (classification task) and variance for numerical labels(regression task)]

force_classification: Whether to force a classification task even if the labels are numerical (typically when labels are integers encoding some feature rather than representing a real cardinal measure) [def: false]

rng: Random Number Generator ((see FIXEDSEED)) [deafult: Random.GLOBAL_RNG]

Notes:

Missing data (in the feature dataset) are supported.

BetaML.Trees.wrap — Function

wrap(node:: DecisionNode, ...)

Called on the root node of a DecsionTree dc in order to add visualization information. In case of a BetaML/DecisionTree this is typically a list of feature names as follows:

wdc = wrap(dc, (featurenames = feature_names, ))