The Bmlt.Utils Module

Bmlt.Utils — Module

Utils module

Provide shared utility functions for various machine learning algorithms. You don't usually need to import from this module, as each other module (Nn, Perceptron, Clusters,...) reexport it.

Module Index

Bmlt.Utils.accuracy

Bmlt.Utils.accuracy

Bmlt.Utils.accuracy

Bmlt.Utils.autoJacobian

Bmlt.Utils.batch

Bmlt.Utils.cosine_distance

Bmlt.Utils.dSoftMax

Bmlt.Utils.getScaleFactors

Bmlt.Utils.l1_distance

Bmlt.Utils.l2_distance

Bmlt.Utils.l2²_distance

Bmlt.Utils.logNormalFixedSd

Bmlt.Utils.lse

Bmlt.Utils.makeMatrix

Bmlt.Utils.normalFixedSd

Bmlt.Utils.oneHotEncoder

Bmlt.Utils.polynomialKernel

Bmlt.Utils.radialKernel

Bmlt.Utils.scale

Bmlt.Utils.sterling

Detailed API

Base.error — Method

Categorical error

Base.error — Method

Categorical error with probabilistic prediction of a single datapoint.

Base.error — Method

Categorical error with probabilistic predictions of a dataset.

Base.reshape — Method

reshape(myNumber, dims..) - Reshape a number as a n dimensional Array

Bmlt.Utils.accuracy — Method

Categorical accuracy

Bmlt.Utils.accuracy — Method

accuracy(x,y;tol)

Categorical accuracy with probabilistic prediction of a single datapoint.

Use the parameter tol [def: 1] to determine the tollerance of the prediction, i.e. if considering "correct" only a prediction where the value with highest probability is the true value (tol = 1), or consider instead the set of tol maximum values.

Bmlt.Utils.accuracy — Method

accuracy(x,y;tol)

Categorical accuracy with probabilistic predictions of a dataset.

Use the parameter tol [def: 1] to determine the tollerance of the prediction, i.e. if considering "correct" only a prediction where the value with highest probability is the true value (tol = 1), or consider instead the set of tol maximum values.

Bmlt.Utils.autoJacobian — Method

autoJacobian(f,x;nY)

Evaluate the Jacobian using AD in the form of a (nY,nX) madrix of first derivatives

Parameters:

f: The function to compute the Jacobian

x: The input to the function where the jacobian has to be computed

nY: The number of outputs of the function f [def: length(f(x))]

Return values:

An Array{Float64,2} of the locally evaluated Jacobian

Notes:

The nY parameter is optional. If provided it avoids having to compute f(x)

Bmlt.Utils.batch — Method

batch(n,bSize;sequential=false)

Return a vector of bSize indeces from 1 to n. Randomly unless the optional parameter sequential is used.

Bmlt.Utils.cosine_distance — Method

Cosine distance

Bmlt.Utils.dSoftMax — Method

dSoftMax(x;β) - Derivative of the softMax function

Bmlt.Utils.getScaleFactors — Method

getScaleFactors(x;skip)

Return the scale factors (for each dimensions) in order to scale a matrix X (n,d) such that each dimension has mean 0 and variance 1.

Parameters

x: the (n × d) dimension matrix to scale on each dimension d

skip: an array of dimension index to skip the scaling [def: []]

Return

A touple whose first elmement is the shift and the second the multiplicative

term to make the scale.

Bmlt.Utils.l1_distance — Method

L1 norm distance (aka Manhattan Distance)

Bmlt.Utils.l2_distance — Method

Euclidean (L2) distance

Bmlt.Utils.l2²_distance — Method

Squared Euclidean (L2) distance

Bmlt.Utils.logNormalFixedSd — Method

log-PDF of a multidimensional normal with no covariance and shared variance across dimensions

Bmlt.Utils.lse — Method

LogSumExp for efficiently computing log(sum(exp.(x)))

Bmlt.Utils.makeMatrix — Method

Transform an Array{T,1} in an Array{T,2} and leave unchanged Array{T,2}.

Bmlt.Utils.normalFixedSd — Method

PDF of a multidimensional normal with no covariance and shared variance across dimensions

Bmlt.Utils.oneHotEncoder — Function

oneHotEncoder(y,d;count)

Encode arrays (or arrays of arrays) of integer data as 0/1 matrices

Parameters:

y: The data to convert (integer, array or array of arrays of integers)

d: The number of dimensions in the output matrik. [def: maximum(maximum.(Y))]

count: Wether to count multiple instances on the same dimension/record or indicate just presence. [def: false]

Bmlt.Utils.polynomialKernel — Method

Polynomial kernel parametrised with c=0 and d=2 (i.e. a quadratic kernel). For other cᵢ and dᵢ use K = (x,y) -> polynomialKernel(x,y,c=cᵢ,d=dᵢ) as kernel function in the supporting algorithms

Bmlt.Utils.radialKernel — Method

Radial Kernel (aka RBF kernel) parametrised with γ=1/2. For other gammas γᵢ use K = (x,y) -> radialKernel(x,y,γ=γᵢ) as kernel function in the supporting algorithms

Bmlt.Utils.scale — Function

scale(x;scalingFactors)

Perform a linear scaling of x using scaling factors scalingFactors.

Parameters

x: the (n × d) dimension matrix to scale on each dimension d

scalingFactors: an tuple of the constant and multiplicative scaling factor

respectively [def: the scaling factors needed to scale x to mean 0 and variance 1]

Return

The scaled matrix

Notes:

also available scale!(x;scalingFactors) for in-place scaling.

Bmlt.Utils.sterling — Method

Sterling number: number of partitions of a set of n elements in k sets